@@ -1,11 +1,14 @@
 import '/backend/backend.dart';
-import '/components/custom_app_bar_widget.dart';
+import '/components/create_post_widget.dart';
+import '/components/custom_appbar_widget.dart';
 import '/flutter_flow/flutter_flow_icon_button.dart';
 import '/flutter_flow/flutter_flow_theme.dart';
 import '/flutter_flow/flutter_flow_toggle_icon.dart';
 import '/flutter_flow/flutter_flow_util.dart';
-import '/flutter_flow/flutter_flow_widgets.dart';
+import '/flutter_flow/random_data_util.dart' as random_data;
+import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:google_fonts/google_fonts.dart';
@@ -14,7 +17,7 @@ import 'community_page_model.dart';
 export 'community_page_model.dart';
 
 class CommunityPageWidget extends StatefulWidget {
-  const CommunityPageWidget({Key? key}) : super(key: key);
+  const CommunityPageWidget({super.key});
 
   @override
   _CommunityPageWidgetState createState() => _CommunityPageWidgetState();
@@ -42,13 +45,23 @@ class _CommunityPageWidgetState extends State<CommunityPageWidget> {
 
   @override
   Widget build(BuildContext context) {
+    if (isiOS) {
+      SystemChrome.setSystemUIOverlayStyle(
+        SystemUiOverlayStyle(
+          statusBarBrightness: Theme.of(context).brightness,
+          systemStatusBarContrastEnforced: true,
+        ),
+      );
+    }
+
     context.watch<FFAppState>();
 
     return GestureDetector(
-      onTap: () => FocusScope.of(context).requestFocus(_model.unfocusNode),
+      onTap: () => _model.unfocusNode.canRequestFocus
+          ? FocusScope.of(context).requestFocus(_model.unfocusNode)
+          : FocusScope.of(context).unfocus(),
       child: Scaffold(
         key: scaffoldKey,
-        backgroundColor: FlutterFlowTheme.of(context).primaryBackground,
         body: SafeArea(
           top: true,
           child: Stack(
@@ -66,18 +79,14 @@ class _CommunityPageWidgetState extends State<CommunityPageWidget> {
                       mainAxisSize: MainAxisSize.max,
                       children: [
                         wrapWithModel(
-                          model: _model.customAppBarModel,
+                          model: _model.customAppbarModel,
                           updateCallback: () => setState(() {}),
-                          child: CustomAppBarWidget(),
-                        ),
-                        Divider(
-                          thickness: 1.0,
-                          color: Color(0xFFACAEAF),
+                          child: const CustomAppbarWidget(),
                         ),
                         Container(
                           width: MediaQuery.sizeOf(context).width * 1.0,
                           height: MediaQuery.sizeOf(context).height * 0.8,
-                          decoration: BoxDecoration(),
+                          decoration: const BoxDecoration(),
                           child: StreamBuilder<List<CommunityTalkRecord>>(
                             stream: queryCommunityTalkRecord(
                               queryBuilder: (communityTalkRecord) =>
@@ -111,12 +120,12 @@ class _CommunityPageWidgetState extends State<CommunityPageWidget> {
                                       listViewCommunityTalkRecordList[
                                           listViewIndex];
                                   return Padding(
-                                    padding: EdgeInsetsDirectional.fromSTEB(
+                                    padding: const EdgeInsetsDirectional.fromSTEB(
                                         10.0, 0.0, 10.0, 0.0),
                                     child: Container(
                                       width: 100.0,
-                                      height: 135.0,
-                                      decoration: BoxDecoration(),
+                                      height: 140.0,
+                                      decoration: const BoxDecoration(),
                                       child: Column(
                                         mainAxisSize: MainAxisSize.max,
                                         children: [
@@ -138,13 +147,11 @@ class _CommunityPageWidgetState extends State<CommunityPageWidget> {
                                                 }.withoutNulls,
                                                 extra: <String, dynamic>{
                                                   kTransitionInfoKey:
-                                                      TransitionInfo(
+                                                      const TransitionInfo(
                                                     hasTransition: true,
                                                     transitionType:
                                                         PageTransitionType
                                                             .bottomToTop,
-                                                    duration: Duration(
-                                                        milliseconds: 300),
                                                   ),
                                                 },
                                               );
@@ -170,31 +177,57 @@ class _CommunityPageWidgetState extends State<CommunityPageWidget> {
                                                       mainAxisAlignment:
                                                           MainAxisAlignment
                                                               .spaceBetween,
+                                                      crossAxisAlignment:
+                                                          CrossAxisAlignment
+                                                              .start,
                                                       children: [
                                                         Container(
-                                                          width: 65.0,
-                                                          height: 20.0,
+                                                          width: 100.0,
+                                                          height: 30.0,
                                                           decoration:
                                                               BoxDecoration(
-                                                            color: Color(
+                                                            color: const Color(
                                                                 0xFFD8D5D5),
+                                                            borderRadius:
+                                                                BorderRadius
+                                                                    .circular(
+                                                                        3.0),
                                                           ),
                                                           child: Align(
                                                             alignment:
-                                                                AlignmentDirectional(
-                                                                    0.00, 0.00),
+                                                                const AlignmentDirectional(
+                                                                    0.0, 0.0),
                                                             child: Text(
                                                               listViewCommunityTalkRecord
                                                                   .category,
+                                                              textAlign:
+                                                                  TextAlign
+                                                                      .justify,
                                                               style: FlutterFlowTheme
                                                                       .of(context)
-                                                                  .bodyMedium,
+                                                                  .bodyMedium
+                                                                  .override(
+                                                                    fontFamily:
+                                                                        FlutterFlowTheme.of(context)
+                                                                            .bodyMediumFamily,
+                                                                    color: const Color(
+                                                                        0xFF272626),
+                                                                    fontSize:
+                                                                        13.0,
+                                                                    fontWeight:
+                                                                        FontWeight
+                                                                            .w300,
+                                                                    useGoogleFonts: GoogleFonts
+                                                                            .asMap()
+                                                                        .containsKey(
+                                                                            FlutterFlowTheme.of(context).bodyMediumFamily),
+                                                                  ),
                                                             ),
                                                           ),
                                                         ),
                                                         Padding(
                                                           padding:
-                                                              EdgeInsetsDirectional
+                                                              const EdgeInsetsDirectional
                                                                   .fromSTEB(
                                                                       0.0,
                                                                       5.0,
@@ -205,7 +238,19 @@ class _CommunityPageWidgetState extends State<CommunityPageWidget> {
                                                                 .title,
                                                             style: FlutterFlowTheme
                                                                     .of(context)
-                                                                .labelMedium,
+                                                                .labelMedium
+                                                                .override(
+                                                                  fontFamily: FlutterFlowTheme.of(
+                                                                          context)
+                                                                      .labelMediumFamily,
+                                                                  fontSize:
+                                                                      16.0,
+                                                                  useGoogleFonts: GoogleFonts
+                                                                          .asMap()
+                                                                      .containsKey(
+                                                                          FlutterFlowTheme.of(context)
+                                                                              .labelMediumFamily),
+                                                                ),
                                                           ),
                                                         ),
                                                       ],
@@ -213,122 +258,178 @@ class _CommunityPageWidgetState extends State<CommunityPageWidget> {
                                                     Column(
                                                       mainAxisSize:
                                                           MainAxisSize.max,
+                                                      mainAxisAlignment:
+                                                          MainAxisAlignment
+                                                              .center,
+                                                      crossAxisAlignment:
+                                                          CrossAxisAlignment
+                                                              .start,
                                                       children: [
+                                                        Container(
+                                                          width:
+                                                              MediaQuery.sizeOf(
+                                                                          context)
+                                                                      .width *
+                                                                  0.65,
+                                                          height:
+                                                              MediaQuery.sizeOf(
+                                                                          context)
+                                                                      .height *
+                                                                  0.02,
+                                                          decoration:
+                                                              const BoxDecoration(),
+                                                          child: Visibility(
+                                                            visible:
+                                                                responsiveVisibility(
+                                                              context: context,
+                                                              tablet: false,
+                                                              tabletLandscape:
+                                                                  false,
+                                                              desktop: false,
+                                                            ),
+                                                            child: AutoSizeText(
+                                                              listViewCommunityTalkRecord
+                                                                  .content,
+                                                              maxLines: 1,
+                                                              style: FlutterFlowTheme
+                                                                      .of(context)
+                                                                  .labelMedium
+                                                                  .override(
+                                                                    fontFamily:
+                                                                        FlutterFlowTheme.of(context)
+                                                                            .labelMediumFamily,
+                                                                    color: FlutterFlowTheme.of(
+                                                                            context)
+                                                                        .accent2,
+                                                                    useGoogleFonts: GoogleFonts
+                                                                            .asMap()
+                                                                        .containsKey(
+                                                                            FlutterFlowTheme.of(context).labelMediumFamily),
+                                                                  ),
+                                                            ),
+                                                          ),
+                                                        ),
                                                         Padding(
                                                           padding:
-                                                              EdgeInsetsDirectional
+                                                              const EdgeInsetsDirectional
                                                                   .fromSTEB(
                                                                       0.0,
+                                                                      8.0,
                                                                       0.0,
-                                                                      0.0,
-                                                                      10.0),
-                                                          child: Text(
-                                                            dateTimeFormat(
-                                                                'jm',
-                                                                listViewCommunityTalkRecord
-                                                                    .uploadTime!),
-                                                            style: FlutterFlowTheme
-                                                                    .of(context)
-                                                                .bodyLarge
-                                                                .override(
-                                                                  fontFamily: FlutterFlowTheme.of(
-                                                                          context)
-                                                                      .bodyLargeFamily,
-                                                                  color: Color(
-                                                                      0xFF8A8888),
-                                                                  fontSize:
-                                                                      11.0,
-                                                                  useGoogleFonts: GoogleFonts
-                                                                          .asMap()
-                                                                      .containsKey(
-                                                                          FlutterFlowTheme.of(context)
-                                                                              .bodyLargeFamily),
+                                                                      0.0),
+                                                          child: Row(
+                                                            mainAxisSize:
+                                                                MainAxisSize
+                                                                    .max,
+                                                            children: [
+                                                              if (responsiveVisibility(
+                                                                context:
+                                                                    context,
+                                                                tablet: false,
+                                                                tabletLandscape:
+                                                                    false,
+                                                                desktop: false,
+                                                              ))
+                                                                Align(
+                                                                  alignment:
+                                                                      const AlignmentDirectional(
+                                                                          -1.0,
+                                                                          0.0),
+                                                                  child: Text(
+                                                                    valueOrDefault<
+                                                                        String>(
+                                                                      listViewCommunityTalkRecord
+                                                                          .address
+                                                                          ?.toString(),
+                                                                      '주소가 없음 ',
+                                                                    ),
+                                                                    style:
+                                                                        const TextStyle(
+                                                                      color: Color(
+                                                                          0xFF797979),
+                                                                      fontSize:
+                                                                          13.0,
+                                                                    ),
+                                                                  ),
                                                                 ),
+                                                              SizedBox(
+                                                                height: 15.0,
+                                                                child:
+                                                                    VerticalDivider(
+                                                                  thickness:
+                                                                      1.0,
+                                                                  color: FlutterFlowTheme.of(
+                                                                          context)
+                                                                      .accent2,
+                                                                ),
+                                                              ),
+                                                              if (responsiveVisibility(
+                                                                context:
+                                                                    context,
+                                                                tablet: false,
+                                                                tabletLandscape:
+                                                                    false,
+                                                                desktop: false,
+                                                              ))
+                                                                Text(
+                                                                  dateTimeFormat(
+                                                                      'relative',
+                                                                      listViewCommunityTalkRecord
+                                                                          .uploadTime!),
+                                                                  style:
+                                                                      const TextStyle(
+                                                                    color: Color(
+                                                                        0xFF797979),
+                                                                    fontSize:
+                                                                        13.0,
+                                                                  ),
+                                                                ),
+                                                            ],
                                                           ),
                                                         ),
                                                       ],
                                                     ),
                                                   ],
                                                 ),
-                                                Column(
-                                                  mainAxisSize:
-                                                      MainAxisSize.max,
-                                                  mainAxisAlignment:
-                                                      MainAxisAlignment.start,
-                                                  crossAxisAlignment:
-                                                      CrossAxisAlignment.start,
-                                                  children: [
-                                                    ClipRRect(
-                                                      borderRadius:
-                                                          BorderRadius.circular(
-                                                              15.0),
-                                                      child: Image.network(
-                                                        listViewCommunityTalkRecord
-                                                            .uploadImage,
-                                                        width: 75.0,
-                                                        height: 75.0,
-                                                        fit: BoxFit.cover,
-                                                      ),
-                                                    ),
-                                                    Row(
+                                                Align(
+                                                  alignment:
+                                                      const AlignmentDirectional(
+                                                          0.0, 0.0),
+                                                  child: Padding(
+                                                    padding:
+                                                        const EdgeInsetsDirectional
+                                                            .fromSTEB(0.0, 10.0,
+                                                                0.0, 0.0),
+                                                    child: Column(
                                                       mainAxisSize:
                                                           MainAxisSize.max,
                                                       mainAxisAlignment:
                                                           MainAxisAlignment
-                                                              .spaceBetween,
+                                                              .start,
+                                                      crossAxisAlignment:
+                                                          CrossAxisAlignment
+                                                              .center,
                                                       children: [
-                                                        Padding(
-                                                          padding:
-                                                              EdgeInsetsDirectional
-                                                                  .fromSTEB(
-                                                                      0.0,
-                                                                      0.0,
-                                                                      5.0,
-                                                                      0.0),
-                                                          child: Row(
-                                                            mainAxisSize:
-                                                                MainAxisSize
-                                                                    .max,
-                                                            mainAxisAlignment:
-                                                                MainAxisAlignment
-                                                                    .center,
-                                                            children: [
-                                                              ToggleIcon(
-                                                                onPressed:
-                                                                    () async {
-                                                                  setState(() => FFAppState()
-                                                                          .likeToggle =
-                                                                      !FFAppState()
-                                                                          .likeToggle);
-                                                                },
-                                                                value: FFAppState()
-                                                                    .likeToggle,
-                                                                onIcon: Icon(
-                                                                  Icons
-                                                                      .favorite_sharp,
-                                                                  color: FlutterFlowTheme.of(
-                                                                          context)
-                                                                      .alternate,
-                                                                  size: 18.0,
-                                                                ),
-                                                                offIcon: Icon(
-                                                                  Icons
-                                                                      .favorite_border,
-                                                                  color: FlutterFlowTheme.of(
-                                                                          context)
-                                                                      .secondaryText,
-                                                                  size: 18.0,
-                                                                ),
-                                                              ),
-                                                              Text(
-                                                                listViewCommunityTalkRecord
-                                                                    .like
-                                                                    .toString(),
-                                                                style: FlutterFlowTheme.of(
-                                                                        context)
-                                                                    .bodyMedium,
-                                                              ),
-                                                            ],
+                                                        ClipRRect(
+                                                          borderRadius:
+                                                              BorderRadius
+                                                                  .circular(
+                                                                      8.0),
+                                                          child: Image.network(
+                                                            listViewCommunityTalkRecord
+                                                                .uploadedImages
+                                                                .first,
+                                                            width: MediaQuery
+                                                                        .sizeOf(
+                                                                            context)
+                                                                    .width *
+                                                                0.2,
+                                                            height: MediaQuery
+                                                                        .sizeOf(
+                                                                            context)
+                                                                    .height *
+                                                                0.08,
+                                                            fit: BoxFit.cover,
                                                           ),
                                                         ),
                                                         Row(
@@ -336,41 +437,171 @@ class _CommunityPageWidgetState extends State<CommunityPageWidget> {
                                                               MainAxisSize.max,
                                                           mainAxisAlignment:
                                                               MainAxisAlignment
-                                                                  .spaceEvenly,
+                                                                  .spaceBetween,
                                                           children: [
-                                                            FaIcon(
-                                                              FontAwesomeIcons
-                                                                  .comment,
-                                                              color: FlutterFlowTheme
-                                                                      .of(context)
-                                                                  .secondaryText,
-                                                              size: 18.0,
-                                                            ),
                                                             Padding(
                                                               padding:
-                                                                  EdgeInsetsDirectional
+                                                                  const EdgeInsetsDirectional
+                                                                      .fromSTEB(
+                                                                          0.0,
+                                                                          0.0,
+                                                                          5.0,
+                                                                          0.0),
+                                                              child: Row(
+                                                                mainAxisSize:
+                                                                    MainAxisSize
+                                                                        .max,
+                                                                mainAxisAlignment:
+                                                                    MainAxisAlignment
+                                                                        .center,
+                                                                children: [
+                                                                  ToggleIcon(
+                                                                    onPressed:
+                                                                        () async {
+                                                                      setState(() => FFAppState()
+                                                                              .likeToggle =
+                                                                          !FFAppState()
+                                                                              .likeToggle);
+                                                                    },
+                                                                    value: FFAppState()
+                                                                        .likeToggle,
+                                                                    onIcon:
+                                                                        Icon(
+                                                                      Icons
+                                                                          .favorite_sharp,
+                                                                      color: FlutterFlowTheme.of(
+                                                                              context)
+                                                                          .alternate,
+                                                                      size:
+                                                                          18.0,
+                                                                    ),
+                                                                    offIcon:
+                                                                        Icon(
+                                                                      Icons
+                                                                          .favorite_border,
+                                                                      color: FlutterFlowTheme.of(
+                                                                              context)
+                                                                          .secondaryText,
+                                                                      size:
+                                                                          18.0,
+                                                                    ),
+                                                                  ),
+                                                                  Text(
+                                                                    listViewCommunityTalkRecord
+                                                                        .like
+                                                                        .toString(),
+                                                                    style: FlutterFlowTheme.of(
+                                                                            context)
+                                                                        .bodyMedium,
+                                                                  ),
+                                                                ],
+                                                              ),
+                                                            ),
+                                                            Row(
+                                                              mainAxisSize:
+                                                                  MainAxisSize
+                                                                      .max,
+                                                              mainAxisAlignment:
+                                                                  MainAxisAlignment
+                                                                      .spaceEvenly,
+                                                              children: [
+                                                                Padding(
+                                                                  padding: const EdgeInsetsDirectional
+                                                                      .fromSTEB(
+                                                                          0.0,
+                                                                          0.0,
+                                                                          10.0,
+                                                                          0.0),
+                                                                  child: FaIcon(
+                                                                    FontAwesomeIcons
+                                                                        .comment,
+                                                                    color: FlutterFlowTheme.of(
+                                                                            context)
+                                                                        .secondaryText,
+                                                                    size: 18.0,
+                                                                  ),
+                                                                ),
+                                                                Padding(
+                                                                  padding: const EdgeInsetsDirectional
                                                                       .fromSTEB(
                                                                           3.0,
                                                                           0.0,
                                                                           0.0,
                                                                           0.0),
-                                                              child: Text(
-                                                                '3',
-                                                                style: FlutterFlowTheme.of(
-                                                                        context)
-                                                                    .bodyMedium,
-                                                              ),
+                                                                  child: StreamBuilder<
+                                                                      List<
+                                                                          CommentRecord>>(
+                                                                    stream:
+                                                                        queryCommentRecord(
+                                                                      queryBuilder:
+                                                                          (commentRecord) =>
+                                                                              commentRecord.where(
+                                                                        'communitTalkRef',
+                                                                        isEqualTo:
+                                                                            listViewCommunityTalkRecord.reference,
+                                                                      ),
+                                                                      singleRecord:
+                                                                          true,
+                                                                    ),
+                                                                    builder:
+                                                                        (context,
+                                                                            snapshot) {
+                                                                      // Customize what your widget looks like when it's loading.
+                                                                      if (!snapshot
+                                                                          .hasData) {
+                                                                        return Center(
+                                                                          child:
+                                                                              SizedBox(
+                                                                            width:
+                                                                                50.0,
+                                                                            height:
+                                                                                50.0,
+                                                                            child:
+                                                                                SpinKitChasingDots(
+                                                                              color: FlutterFlowTheme.of(context).info,
+                                                                              size: 50.0,
+                                                                            ),
+                                                                          ),
+                                                                        );
+                                                                      }
+                                                                      List<CommentRecord>
+                                                                          textCommentRecordList =
+                                                                          snapshot
+                                                                              .data!;
+                                                                      // Return an empty Container when the item does not exist.
+                                                                      if (snapshot
+                                                                          .data!
+                                                                          .isEmpty) {
+                                                                        return Container();
+                                                                      }
+                                                                      final textCommentRecord = textCommentRecordList
+                                                                              .isNotEmpty
+                                                                          ? textCommentRecordList
+                                                                              .first
+                                                                          : null;
+                                                                      return Text(
+                                                                        random_data
+                                                                            .randomInteger(0,
+                                                                                10)
+                                                                            .toString(),
+                                                                        style: FlutterFlowTheme.of(context)
+                                                                            .bodyMedium,
+                                                                      );
+                                                                    },
+                                                                  ),
+                                                                ),
+                                                              ],
                                                             ),
                                                           ],
                                                         ),
                                                       ],
                                                     ),
-                                                  ],
+                                                  ),
                                                 ),
                                               ],
                                             ),
                                           ),
-                                          Divider(
+                                          const Divider(
                                             thickness: 1.0,
                                             color: Color(0xFFB7B0B0),
                                           ),
@@ -389,29 +620,40 @@ class _CommunityPageWidgetState extends State<CommunityPageWidget> {
                 ],
               ),
               Align(
-                alignment: AlignmentDirectional(0.73, 0.86),
+                alignment: const AlignmentDirectional(0.88, 0.91),
                 child: FlutterFlowIconButton(
                   borderColor: Colors.transparent,
                   borderRadius: 27.5,
                   borderWidth: 1.0,
                   buttonSize: 55.0,
-                  fillColor: FlutterFlowTheme.of(context).main2,
+                  fillColor: const Color(0xFFE15A19),
                   icon: Icon(
                     Icons.add,
                     color: FlutterFlowTheme.of(context).primaryText,
                     size: 24.0,
                   ),
                   onPressed: () async {
-                    context.pushNamed(
-                      'CUploadPage',
-                      extra: <String, dynamic>{
-                        kTransitionInfoKey: TransitionInfo(
-                          hasTransition: true,
-                          transitionType: PageTransitionType.bottomToTop,
-                          duration: Duration(milliseconds: 300),
-                        ),
+                    await showModalBottomSheet(
+                      isScrollControlled: true,
+                      backgroundColor: const Color(0x44000000),
+                      enableDrag: false,
+                      context: context,
+                      builder: (context) {
+                        return GestureDetector(
+                          onTap: () => _model.unfocusNode.canRequestFocus
+                              ? FocusScope.of(context)
+                                  .requestFocus(_model.unfocusNode)
+                              : FocusScope.of(context).unfocus(),
+                          child: Padding(
+                            padding: MediaQuery.viewInsetsOf(context),
+                            child: SizedBox(
+                              height: MediaQuery.sizeOf(context).height * 1.0,
+                              child: const CreatePostWidget(),
+                            ),
+                          ),
+                        );
                       },
-                    );
+                    ).then((value) => safeSetState(() {}));
                   },
                 ),
               ),
