@@ -2,25 +2,37 @@ import '/auth/firebase_auth/auth_util.dart';
 import '/backend/backend.dart';
 import '/backend/firebase_storage/storage.dart';
 import '/components/failed_popup_widget.dart';
-import '/components/sucess_pop_up_widget.dart';
+import '/flutter_flow/flutter_flow_choice_chips.dart';
 import '/flutter_flow/flutter_flow_drop_down.dart';
 import '/flutter_flow/flutter_flow_icon_button.dart';
 import '/flutter_flow/flutter_flow_theme.dart';
 import '/flutter_flow/flutter_flow_util.dart';
-import '/flutter_flow/flutter_flow_widgets.dart';
 import '/flutter_flow/form_field_controller.dart';
 import '/flutter_flow/upload_data.dart';
-import 'package:cloud_firestore/cloud_firestore.dart';
+import '/pages/sucess_pop_up/sucess_pop_up_widget.dart';
 import 'package:easy_debounce/easy_debounce.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_spinkit/flutter_spinkit.dart';
+import 'package:flutter/services.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
 import 'upload_post_model.dart';
 export 'upload_post_model.dart';
 
 class UploadPostWidget extends StatefulWidget {
-  const UploadPostWidget({Key? key}) : super(key: key);
+  const UploadPostWidget({
+    super.key,
+    this.location,
+    this.category,
+    this.title,
+    this.price,
+    this.desc,
+  });
+
+  final LatLng? location;
+  final String? category;
+  final String? title;
+  final int? price;
+  final String? desc;
 
   @override
   _UploadPostWidgetState createState() => _UploadPostWidgetState();
@@ -36,8 +48,16 @@ class _UploadPostWidgetState extends State<UploadPostWidget> {
     super.initState();
     _model = createModel(context, () => UploadPostModel());
 
-    _model.textController1 ??= TextEditingController();
-    _model.textController2 ??= TextEditingController();
+    _model.textController1 ??= TextEditingController(text: widget.title);
+    _model.textFieldFocusNode1 ??= FocusNode();
+
+    _model.textController2 ??=
+        TextEditingController(text: widget.price?.toString());
+    _model.textFieldFocusNode2 ??= FocusNode();
+
+    _model.textController3 ??= TextEditingController(text: widget.desc);
+    _model.textFieldFocusNode3 ??= FocusNode();
+
     WidgetsBinding.instance.addPostFrameCallback((_) => setState(() {}));
   }
 
@@ -50,11 +70,19 @@ class _UploadPostWidgetState extends State<UploadPostWidget> {
 
   @override
   Widget build(BuildContext context) {
+    if (isiOS) {
+      SystemChrome.setSystemUIOverlayStyle(
+        SystemUiOverlayStyle(
+          statusBarBrightness: Theme.of(context).brightness,
+          systemStatusBarContrastEnforced: true,
+        ),
+      );
+    }
+
     context.watch<FFAppState>();
 
     return Scaffold(
       key: scaffoldKey,
-      backgroundColor: Colors.white,
       body: SafeArea(
         top: true,
         child: Column(
@@ -63,14 +91,16 @@ class _UploadPostWidgetState extends State<UploadPostWidget> {
           children: [
             Container(
               width: MediaQuery.sizeOf(context).width * 1.0,
-              height: 50.0,
-              decoration: BoxDecoration(),
+              height: 65.0,
+              decoration: BoxDecoration(
+                color: FlutterFlowTheme.of(context).main1,
+              ),
               child: Row(
                 mainAxisSize: MainAxisSize.max,
                 children: [
                   Padding(
                     padding:
-                        EdgeInsetsDirectional.fromSTEB(15.0, 0.0, 7.0, 0.0),
+                        const EdgeInsetsDirectional.fromSTEB(15.0, 0.0, 7.0, 0.0),
                     child: InkWell(
                       splashColor: Colors.transparent,
                       focusColor: Colors.transparent,
@@ -81,346 +111,685 @@ class _UploadPostWidgetState extends State<UploadPostWidget> {
                       },
                       child: Icon(
                         Icons.arrow_back_ios,
-                        color: FlutterFlowTheme.of(context).sub1,
+                        color: FlutterFlowTheme.of(context).primaryText,
                         size: 24.0,
                       ),
                     ),
                   ),
-                  Text(
-                    'New Listing ',
-                    style: FlutterFlowTheme.of(context).titleSmall,
+                  Align(
+                    alignment: const AlignmentDirectional(0.0, 0.0),
+                    child: Padding(
+                      padding:
+                          const EdgeInsetsDirectional.fromSTEB(95.0, 0.0, 0.0, 0.0),
+                      child: Text(
+                        '거래를 시작하세요',
+                        textAlign: TextAlign.center,
+                        style:
+                            FlutterFlowTheme.of(context).titleMedium.override(
+                                  fontFamily: FlutterFlowTheme.of(context)
+                                      .titleMediumFamily,
+                                  color: FlutterFlowTheme.of(context).sub1,
+                                  fontWeight: FontWeight.w600,
+                                  useGoogleFonts: GoogleFonts.asMap()
+                                      .containsKey(FlutterFlowTheme.of(context)
+                                          .titleMediumFamily),
+                                ),
+                      ),
+                    ),
                   ),
                 ],
               ),
             ),
             Padding(
-              padding: EdgeInsetsDirectional.fromSTEB(20.0, 0.0, 20.0, 0.0),
+              padding: const EdgeInsetsDirectional.fromSTEB(20.0, 0.0, 0.0, 0.0),
+              child: Column(
+                mainAxisSize: MainAxisSize.max,
+                children: [
+                  Align(
+                    alignment: const AlignmentDirectional(-1.0, 0.0),
+                    child: Padding(
+                      padding:
+                          const EdgeInsetsDirectional.fromSTEB(0.0, 0.0, 0.0, 10.0),
+                      child: Text(
+                        'where to meet',
+                        style:
+                            FlutterFlowTheme.of(context).titleMedium.override(
+                                  fontFamily: FlutterFlowTheme.of(context)
+                                      .titleMediumFamily,
+                                  fontSize: 16.0,
+                                  useGoogleFonts: GoogleFonts.asMap()
+                                      .containsKey(FlutterFlowTheme.of(context)
+                                          .titleMediumFamily),
+                                ),
+                      ),
+                    ),
+                  ),
+                  Padding(
+                    padding:
+                        const EdgeInsetsDirectional.fromSTEB(0.0, 0.0, 20.0, 10.0),
+                    child: InkWell(
+                      splashColor: Colors.transparent,
+                      focusColor: Colors.transparent,
+                      hoverColor: Colors.transparent,
+                      highlightColor: Colors.transparent,
+                      onTap: () async {
+                        context.pushNamed(
+                          'mappageCopy',
+                          queryParameters: {
+                            'category': serializeParam(
+                              _model.dropDownValue,
+                              ParamType.String,
+                            ),
+                            'title': serializeParam(
+                              _model.textController1.text,
+                              ParamType.String,
+                            ),
+                            'price': serializeParam(
+                              int.tryParse(_model.textController2.text),
+                              ParamType.int,
+                            ),
+                            'description': serializeParam(
+                              _model.textController3.text,
+                              ParamType.String,
+                            ),
+                            'negoitable': serializeParam(
+                              _model.checkboxValue,
+                              ParamType.bool,
+                            ),
+                          }.withoutNulls,
+                        );
+                      },
+                      child: Container(
+                        width: MediaQuery.sizeOf(context).width * 1.0,
+                        height: MediaQuery.sizeOf(context).height * 0.05,
+                        decoration: BoxDecoration(
+                          borderRadius: BorderRadius.circular(7.0),
+                          border: Border.all(
+                            color: const Color(0xFFB8B6B6),
+                            width: 2.0,
+                          ),
+                        ),
+                        alignment: const AlignmentDirectional(0.0, 0.0),
+                        child: Padding(
+                          padding: const EdgeInsetsDirectional.fromSTEB(
+                              12.0, 0.0, 12.0, 0.0),
+                          child: Row(
+                            mainAxisSize: MainAxisSize.max,
+                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                            children: [
+                              Text(
+                                'Select',
+                                style: FlutterFlowTheme.of(context)
+                                    .labelMedium
+                                    .override(
+                                      fontFamily: FlutterFlowTheme.of(context)
+                                          .labelMediumFamily,
+                                      color: const Color(0xFFA3A2A2),
+                                      useGoogleFonts: GoogleFonts.asMap()
+                                          .containsKey(
+                                              FlutterFlowTheme.of(context)
+                                                  .labelMediumFamily),
+                                    ),
+                              ),
+                              const Icon(
+                                Icons.arrow_forward_ios,
+                                color: Color(0xFFA3A2A2),
+                                size: 24.0,
+                              ),
+                            ],
+                          ),
+                        ),
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ),
+            Row(
+              mainAxisSize: MainAxisSize.max,
+              children: [
+                Padding(
+                  padding: const EdgeInsetsDirectional.fromSTEB(0.0, 0.0, 12.0, 0.0),
+                  child: FlutterFlowIconButton(
+                    borderColor: Colors.transparent,
+                    borderRadius: 30.0,
+                    borderWidth: 1.0,
+                    buttonSize: 60.0,
+                    icon: Icon(
+                      Icons.add_photo_alternate_outlined,
+                      color: FlutterFlowTheme.of(context).primary,
+                      size: 32.0,
+                    ),
+                    showLoadingIndicator: true,
+                    onPressed: () async {
+                      // uploadPhoto
+                      final selectedMedia = await selectMedia(
+                        mediaSource: MediaSource.photoGallery,
+                        multiImage: true,
+                      );
+                      if (selectedMedia != null &&
+                          selectedMedia.every((m) =>
+                              validateFileFormat(m.storagePath, context))) {
+                        setState(() => _model.isDataUploading = true);
+                        var selectedUploadedFiles = <FFUploadedFile>[];
+
+                        var downloadUrls = <String>[];
+                        try {
+                          selectedUploadedFiles = selectedMedia
+                              .map((m) => FFUploadedFile(
+                                    name: m.storagePath.split('/').last,
+                                    bytes: m.bytes,
+                                    height: m.dimensions?.height,
+                                    width: m.dimensions?.width,
+                                    blurHash: m.blurHash,
+                                  ))
+                              .toList();
+
+                          downloadUrls = (await Future.wait(
+                            selectedMedia.map(
+                              (m) async =>
+                                  await uploadData(m.storagePath, m.bytes),
+                            ),
+                          ))
+                              .where((u) => u != null)
+                              .map((u) => u!)
+                              .toList();
+                        } finally {
+                          _model.isDataUploading = false;
+                        }
+                        if (selectedUploadedFiles.length ==
+                                selectedMedia.length &&
+                            downloadUrls.length == selectedMedia.length) {
+                          setState(() {
+                            _model.uploadedLocalFiles = selectedUploadedFiles;
+                            _model.uploadedFileUrls = downloadUrls;
+                          });
+                        } else {
+                          setState(() {});
+                          return;
+                        }
+                      }
+                    },
+                  ),
+                ),
+                Container(
+                  width: 270.0,
+                  height: 50.0,
+                  decoration: const BoxDecoration(),
+                  child: Builder(
+                    builder: (context) {
+                      final uploadimage = _model.uploadedFileUrls.toList();
+                      return ListView.builder(
+                        padding: EdgeInsets.zero,
+                        scrollDirection: Axis.horizontal,
+                        itemCount: uploadimage.length,
+                        itemBuilder: (context, uploadimageIndex) {
+                          final uploadimageItem = uploadimage[uploadimageIndex];
+                          return Padding(
+                            padding: const EdgeInsetsDirectional.fromSTEB(
+                                7.0, 0.0, 0.0, 0.0),
+                            child: Container(
+                              width: 50.0,
+                              decoration: BoxDecoration(
+                                borderRadius: BorderRadius.circular(10.0),
+                                border: Border.all(
+                                  color: const Color(0xFFA8A2A2),
+                                  width: 2.0,
+                                ),
+                              ),
+                              alignment: const AlignmentDirectional(0.0, 0.0),
+                              child: ClipRRect(
+                                borderRadius: BorderRadius.circular(8.0),
+                                child: Image.network(
+                                  uploadimageItem,
+                                  width: 300.0,
+                                  height: 200.0,
+                                  fit: BoxFit.cover,
+                                ),
+                              ),
+                            ),
+                          );
+                        },
+                      );
+                    },
+                  ),
+                ),
+              ].addToStart(const SizedBox(width: 5.0)),
+            ),
+            Padding(
+              padding: const EdgeInsetsDirectional.fromSTEB(20.0, 0.0, 0.0, 0.0),
+              child: Column(
+                mainAxisSize: MainAxisSize.max,
+                children: [
+                  Padding(
+                    padding:
+                        const EdgeInsetsDirectional.fromSTEB(0.0, 0.0, 20.0, 10.0),
+                    child: FlutterFlowDropDown<String>(
+                      controller: _model.dropDownValueController ??=
+                          FormFieldController<String>(
+                        _model.dropDownValue ??=
+                            widget.category != null && widget.category != ''
+                                ? widget.category
+                                : '',
+                      ),
+                      options: const ['자유', '후기', '질문', '정보', '공유', '기타', '전체'],
+                      onChanged: (val) =>
+                          setState(() => _model.dropDownValue = val),
+                      width: MediaQuery.sizeOf(context).width * 1.0,
+                      height: 50.0,
+                      textStyle: FlutterFlowTheme.of(context)
+                          .labelLarge
+                          .override(
+                            fontFamily:
+                                FlutterFlowTheme.of(context).labelLargeFamily,
+                            color: FlutterFlowTheme.of(context).secondaryText,
+                            fontWeight: FontWeight.normal,
+                            useGoogleFonts: GoogleFonts.asMap().containsKey(
+                                FlutterFlowTheme.of(context).labelLargeFamily),
+                          ),
+                      hintText: 'Category',
+                      icon: Icon(
+                        Icons.keyboard_arrow_down_rounded,
+                        color: FlutterFlowTheme.of(context).secondaryText,
+                        size: 24.0,
+                      ),
+                      fillColor:
+                          FlutterFlowTheme.of(context).secondaryBackground,
+                      elevation: 2.0,
+                      borderColor: const Color(0xFFB8B6B6),
+                      borderWidth: 2.0,
+                      borderRadius: 8.0,
+                      margin:
+                          const EdgeInsetsDirectional.fromSTEB(16.0, 4.0, 16.0, 4.0),
+                      hidesUnderline: true,
+                      isSearchable: false,
+                      isMultiSelect: false,
+                    ),
+                  ),
+                  Align(
+                    alignment: const AlignmentDirectional(-1.0, 0.0),
+                    child: Padding(
+                      padding:
+                          const EdgeInsetsDirectional.fromSTEB(0.0, 0.0, 0.0, 10.0),
+                      child: Text(
+                        'Title',
+                        style:
+                            FlutterFlowTheme.of(context).titleMedium.override(
+                                  fontFamily: FlutterFlowTheme.of(context)
+                                      .titleMediumFamily,
+                                  fontSize: 16.0,
+                                  useGoogleFonts: GoogleFonts.asMap()
+                                      .containsKey(FlutterFlowTheme.of(context)
+                                          .titleMediumFamily),
+                                ),
+                      ),
+                    ),
+                  ),
+                  Padding(
+                    padding:
+                        const EdgeInsetsDirectional.fromSTEB(0.0, 0.0, 20.0, 10.0),
+                    child: Container(
+                      width: MediaQuery.sizeOf(context).width * 1.0,
+                      height: MediaQuery.sizeOf(context).height * 0.05,
+                      decoration: BoxDecoration(
+                        borderRadius: BorderRadius.circular(7.0),
+                        border: Border.all(
+                          color: const Color(0xFFB8B6B6),
+                          width: 2.0,
+                        ),
+                      ),
+                      alignment: const AlignmentDirectional(0.0, 0.0),
+                      child: TextFormField(
+                        controller: _model.textController1,
+                        focusNode: _model.textFieldFocusNode1,
+                        onChanged: (_) => EasyDebounce.debounce(
+                          '_model.textController1',
+                          const Duration(milliseconds: 2000),
+                          () => setState(() {}),
+                        ),
+                        textCapitalization: TextCapitalization.sentences,
+                        obscureText: false,
+                        decoration: InputDecoration(
+                          hintText: 'Title',
+                          hintStyle: FlutterFlowTheme.of(context)
+                              .labelLarge
+                              .override(
+                                fontFamily: FlutterFlowTheme.of(context)
+                                    .labelLargeFamily,
+                                color: const Color(0xFFA3A2A2),
+                                useGoogleFonts: GoogleFonts.asMap().containsKey(
+                                    FlutterFlowTheme.of(context)
+                                        .labelLargeFamily),
+                              ),
+                          enabledBorder: InputBorder.none,
+                          focusedBorder: InputBorder.none,
+                          errorBorder: InputBorder.none,
+                          focusedErrorBorder: InputBorder.none,
+                          contentPadding: const EdgeInsetsDirectional.fromSTEB(
+                              16.0, 0.0, 16.0, 0.0),
+                        ),
+                        style: FlutterFlowTheme.of(context).bodyLarge,
+                        textAlign: TextAlign.start,
+                        maxLines: 2,
+                        cursorColor: FlutterFlowTheme.of(context).backG1,
+                        validator: _model.textController1Validator
+                            .asValidator(context),
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsetsDirectional.fromSTEB(20.0, 0.0, 0.0, 0.0),
+              child: Column(
+                mainAxisSize: MainAxisSize.max,
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Align(
+                    alignment: const AlignmentDirectional(-1.0, 0.0),
+                    child: Padding(
+                      padding:
+                          const EdgeInsetsDirectional.fromSTEB(0.0, 0.0, 0.0, 10.0),
+                      child: Text(
+                        'Listing type',
+                        style:
+                            FlutterFlowTheme.of(context).titleMedium.override(
+                                  fontFamily: FlutterFlowTheme.of(context)
+                                      .titleMediumFamily,
+                                  fontSize: 16.0,
+                                  useGoogleFonts: GoogleFonts.asMap()
+                                      .containsKey(FlutterFlowTheme.of(context)
+                                          .titleMediumFamily),
+                                ),
+                      ),
+                    ),
+                  ),
+                  Padding(
+                    padding:
+                        const EdgeInsetsDirectional.fromSTEB(0.0, 0.0, 0.0, 10.0),
+                    child: FlutterFlowChoiceChips(
+                      options: const [ChipData('For sale'), ChipData('Free')],
+                      onChanged: (val) =>
+                          setState(() => _model.choiceChipsValue = val?.first),
+                      selectedChipStyle: ChipStyle(
+                        backgroundColor: FlutterFlowTheme.of(context).sub1,
+                        textStyle: FlutterFlowTheme.of(context)
+                            .bodyMedium
+                            .override(
+                              fontFamily:
+                                  FlutterFlowTheme.of(context).bodyMediumFamily,
+                              color: FlutterFlowTheme.of(context).primaryText,
+                              useGoogleFonts: GoogleFonts.asMap().containsKey(
+                                  FlutterFlowTheme.of(context)
+                                      .bodyMediumFamily),
+                            ),
+                        iconColor: FlutterFlowTheme.of(context).primaryText,
+                        iconSize: 0.0,
+                        elevation: 4.0,
+                        borderRadius: BorderRadius.circular(16.0),
+                      ),
+                      unselectedChipStyle: ChipStyle(
+                        backgroundColor: const Color(0x00000000),
+                        textStyle: FlutterFlowTheme.of(context)
+                            .bodyMedium
+                            .override(
+                              fontFamily:
+                                  FlutterFlowTheme.of(context).bodyMediumFamily,
+                              color: FlutterFlowTheme.of(context).secondaryText,
+                              useGoogleFonts: GoogleFonts.asMap().containsKey(
+                                  FlutterFlowTheme.of(context)
+                                      .bodyMediumFamily),
+                            ),
+                        iconColor: FlutterFlowTheme.of(context).secondaryText,
+                        iconSize: 18.0,
+                        elevation: 0.0,
+                        borderRadius: BorderRadius.circular(16.0),
+                      ),
+                      chipSpacing: 12.0,
+                      rowSpacing: 12.0,
+                      multiselect: false,
+                      initialized: _model.choiceChipsValue != null,
+                      alignment: WrapAlignment.center,
+                      controller: _model.choiceChipsValueController ??=
+                          FormFieldController<List<String>>(
+                        ['For sale'],
+                      ),
+                      wrapped: true,
+                    ),
+                  ),
+                  Padding(
+                    padding:
+                        const EdgeInsetsDirectional.fromSTEB(0.0, 0.0, 20.0, 0.0),
+                    child: Container(
+                      width: MediaQuery.sizeOf(context).width * 1.0,
+                      height: MediaQuery.sizeOf(context).height * 0.055,
+                      decoration: BoxDecoration(
+                        borderRadius: BorderRadius.circular(7.0),
+                        border: Border.all(
+                          color: const Color(0xFFB8B6B6),
+                          width: 2.0,
+                        ),
+                      ),
+                      alignment: const AlignmentDirectional(0.0, 0.0),
+                      child: Align(
+                        alignment: const AlignmentDirectional(0.0, 0.0),
+                        child: TextFormField(
+                          controller: _model.textController2,
+                          focusNode: _model.textFieldFocusNode2,
+                          onChanged: (_) => EasyDebounce.debounce(
+                            '_model.textController2',
+                            const Duration(milliseconds: 2000),
+                            () => setState(() {}),
+                          ),
+                          textCapitalization: TextCapitalization.sentences,
+                          obscureText: false,
+                          decoration: InputDecoration(
+                            hintText: '\$ Price',
+                            hintStyle: FlutterFlowTheme.of(context)
+                                .labelLarge
+                                .override(
+                                  fontFamily: FlutterFlowTheme.of(context)
+                                      .labelLargeFamily,
+                                  color: const Color(0xFFA3A2A2),
+                                  useGoogleFonts: GoogleFonts.asMap()
+                                      .containsKey(FlutterFlowTheme.of(context)
+                                          .labelLargeFamily),
+                                ),
+                            enabledBorder: InputBorder.none,
+                            focusedBorder: InputBorder.none,
+                            errorBorder: InputBorder.none,
+                            focusedErrorBorder: InputBorder.none,
+                            contentPadding: const EdgeInsetsDirectional.fromSTEB(
+                                16.0, 0.0, 16.0, 0.0),
+                          ),
+                          style: FlutterFlowTheme.of(context).bodyLarge,
+                          textAlign: TextAlign.start,
+                          maxLines: null,
+                          keyboardType: TextInputType.number,
+                          cursorColor: FlutterFlowTheme.of(context).backG1,
+                          validator: _model.textController2Validator
+                              .asValidator(context),
+                        ),
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsetsDirectional.fromSTEB(20.0, 10.0, 0.0, 1.0),
+              child: Row(
+                mainAxisSize: MainAxisSize.max,
+                children: [
+                  Container(
+                    width: 22.0,
+                    height: 22.0,
+                    decoration: BoxDecoration(
+                      borderRadius: BorderRadius.circular(2.0),
+                      border: Border.all(
+                        color: const Color(0xFFA1A0A0),
+                        width: 1.0,
+                      ),
+                    ),
+                    child: Theme(
+                      data: ThemeData(
+                        checkboxTheme: CheckboxThemeData(
+                          visualDensity: VisualDensity.compact,
+                          materialTapTargetSize:
+                              MaterialTapTargetSize.shrinkWrap,
+                          shape: RoundedRectangleBorder(
+                            borderRadius: BorderRadius.circular(4.0),
+                          ),
+                        ),
+                        unselectedWidgetColor:
+                            FlutterFlowTheme.of(context).primaryBackground,
+                      ),
+                      child: Checkbox(
+                        value: _model.checkboxValue ??= false,
+                        onChanged: (newValue) async {
+                          setState(() => _model.checkboxValue = newValue!);
+                        },
+                        activeColor: FlutterFlowTheme.of(context).sub1,
+                        checkColor: FlutterFlowTheme.of(context).primaryText,
+                      ),
+                    ),
+                  ),
+                  Padding(
+                    padding:
+                        const EdgeInsetsDirectional.fromSTEB(10.0, 0.0, 0.0, 0.0),
+                    child: Text(
+                      'Open to offers',
+                      style: FlutterFlowTheme.of(context).titleSmall,
+                    ),
+                  ),
+                ],
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsetsDirectional.fromSTEB(20.0, 0.0, 20.0, 0.0),
               child: Container(
                 width: MediaQuery.sizeOf(context).width * 1.0,
                 decoration: BoxDecoration(
                   color: FlutterFlowTheme.of(context).secondaryBackground,
                 ),
-                child: Padding(
-                  padding: EdgeInsetsDirectional.fromSTEB(0.0, 20.0, 0.0, 0.0),
-                  child: FlutterFlowDropDown<String>(
-                    controller: _model.dropDownValueController ??=
-                        FormFieldController<String>(null),
-                    options: ['자유', '후기', '질문', '정보', '공유', '기타', '전체'],
-                    onChanged: (val) =>
-                        setState(() => _model.dropDownValue = val),
-                    width: 300.0,
-                    height: 50.0,
-                    textStyle: FlutterFlowTheme.of(context).bodyMedium.override(
-                          fontFamily:
-                              FlutterFlowTheme.of(context).bodyMediumFamily,
-                          fontSize: 14.0,
-                          fontWeight: FontWeight.w500,
-                          useGoogleFonts: GoogleFonts.asMap().containsKey(
-                              FlutterFlowTheme.of(context).bodyMediumFamily),
-                        ),
-                    hintText: 'Category',
-                    icon: Icon(
-                      Icons.keyboard_arrow_down_rounded,
-                      color: FlutterFlowTheme.of(context).secondaryText,
-                      size: 24.0,
-                    ),
-                    fillColor: FlutterFlowTheme.of(context).secondaryBackground,
-                    elevation: 2.0,
-                    borderColor: FlutterFlowTheme.of(context).backG1,
-                    borderWidth: 2.0,
-                    borderRadius: 8.0,
-                    margin:
-                        EdgeInsetsDirectional.fromSTEB(16.0, 4.0, 16.0, 4.0),
-                    hidesUnderline: true,
-                    isSearchable: false,
-                    isMultiSelect: false,
-                  ),
-                ),
               ),
             ),
             Container(
               width: double.infinity,
-              decoration: BoxDecoration(),
-              child: Padding(
-                padding: EdgeInsetsDirectional.fromSTEB(20.0, 0.0, 20.0, 0.0),
-                child: TextFormField(
-                  controller: _model.textController1,
-                  onChanged: (_) => EasyDebounce.debounce(
-                    '_model.textController1',
-                    Duration(milliseconds: 2000),
-                    () => setState(() {}),
-                  ),
-                  autofocus: true,
-                  textCapitalization: TextCapitalization.sentences,
-                  obscureText: false,
-                  decoration: InputDecoration(
-                    hintText: 'What\'s happening?',
-                    hintStyle: FlutterFlowTheme.of(context).labelLarge,
-                    enabledBorder: OutlineInputBorder(
-                      borderSide: BorderSide(
-                        color: FlutterFlowTheme.of(context).backG1,
-                        width: 2.0,
-                      ),
-                      borderRadius: const BorderRadius.only(
-                        topLeft: Radius.circular(4.0),
-                        topRight: Radius.circular(4.0),
-                      ),
-                    ),
-                    focusedBorder: OutlineInputBorder(
-                      borderSide: BorderSide(
-                        color: FlutterFlowTheme.of(context).primary,
-                        width: 2.0,
-                      ),
-                      borderRadius: const BorderRadius.only(
-                        topLeft: Radius.circular(4.0),
-                        topRight: Radius.circular(4.0),
-                      ),
-                    ),
-                    errorBorder: OutlineInputBorder(
-                      borderSide: BorderSide(
-                        color: FlutterFlowTheme.of(context).error,
-                        width: 2.0,
-                      ),
-                      borderRadius: const BorderRadius.only(
-                        topLeft: Radius.circular(4.0),
-                        topRight: Radius.circular(4.0),
-                      ),
-                    ),
-                    focusedErrorBorder: OutlineInputBorder(
-                      borderSide: BorderSide(
-                        color: FlutterFlowTheme.of(context).error,
-                        width: 2.0,
-                      ),
-                      borderRadius: const BorderRadius.only(
-                        topLeft: Radius.circular(4.0),
-                        topRight: Radius.circular(4.0),
-                      ),
-                    ),
-                    contentPadding:
-                        EdgeInsetsDirectional.fromSTEB(16.0, 8.0, 16.0, 12.0),
-                  ),
-                  style: FlutterFlowTheme.of(context).bodyLarge,
-                  textAlign: TextAlign.start,
-                  maxLines: 7,
-                  cursorColor: FlutterFlowTheme.of(context).backG1,
-                  validator:
-                      _model.textController1Validator.asValidator(context),
-                ),
-              ),
+              decoration: const BoxDecoration(),
             ),
             Padding(
-              padding: EdgeInsetsDirectional.fromSTEB(0.0, 0.0, 0.0, 60.0),
-              child: Container(
-                width: double.infinity,
-                height: 50.0,
-                decoration: BoxDecoration(),
-                child: Padding(
-                  padding: EdgeInsetsDirectional.fromSTEB(20.0, 0.0, 20.0, 0.0),
-                  child: TextFormField(
-                    controller: _model.textController2,
-                    onChanged: (_) => EasyDebounce.debounce(
-                      '_model.textController2',
-                      Duration(milliseconds: 2000),
-                      () => setState(() {}),
-                    ),
-                    autofocus: true,
-                    textCapitalization: TextCapitalization.sentences,
-                    obscureText: false,
-                    decoration: InputDecoration(
-                      hintText: 'Title for your post',
-                      hintStyle: FlutterFlowTheme.of(context).labelLarge,
-                      enabledBorder: OutlineInputBorder(
-                        borderSide: BorderSide(
-                          color: FlutterFlowTheme.of(context).backG1,
-                          width: 2.0,
-                        ),
-                        borderRadius: const BorderRadius.only(
-                          topLeft: Radius.circular(4.0),
-                          topRight: Radius.circular(4.0),
-                        ),
-                      ),
-                      focusedBorder: OutlineInputBorder(
-                        borderSide: BorderSide(
-                          color: FlutterFlowTheme.of(context).primary,
-                          width: 2.0,
-                        ),
-                        borderRadius: const BorderRadius.only(
-                          topLeft: Radius.circular(4.0),
-                          topRight: Radius.circular(4.0),
-                        ),
-                      ),
-                      errorBorder: OutlineInputBorder(
-                        borderSide: BorderSide(
-                          color: FlutterFlowTheme.of(context).error,
-                          width: 2.0,
-                        ),
-                        borderRadius: const BorderRadius.only(
-                          topLeft: Radius.circular(4.0),
-                          topRight: Radius.circular(4.0),
-                        ),
-                      ),
-                      focusedErrorBorder: OutlineInputBorder(
-                        borderSide: BorderSide(
-                          color: FlutterFlowTheme.of(context).error,
-                          width: 2.0,
-                        ),
-                        borderRadius: const BorderRadius.only(
-                          topLeft: Radius.circular(4.0),
-                          topRight: Radius.circular(4.0),
-                        ),
-                      ),
-                      contentPadding:
-                          EdgeInsetsDirectional.fromSTEB(16.0, 8.0, 16.0, 12.0),
-                    ),
-                    style: FlutterFlowTheme.of(context).bodyLarge,
-                    textAlign: TextAlign.start,
-                    maxLines: 7,
-                    cursorColor: FlutterFlowTheme.of(context).backG1,
-                    validator:
-                        _model.textController2Validator.asValidator(context),
-                  ),
-                ),
-              ),
-            ),
-            Builder(
-              builder: (context) {
-                final uploadedImages = _model.uploadedFileUrls.toList();
-                return SingleChildScrollView(
-                  scrollDirection: Axis.horizontal,
-                  child: Row(
-                    mainAxisSize: MainAxisSize.max,
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: List.generate(uploadedImages.length,
-                        (uploadedImagesIndex) {
-                      final uploadedImagesItem =
-                          uploadedImages[uploadedImagesIndex];
-                      return Padding(
-                        padding:
-                            EdgeInsetsDirectional.fromSTEB(7.0, 0.0, 0.0, 0.0),
-                        child: Container(
-                          width: 100.0,
-                          height: 100.0,
-                          decoration: BoxDecoration(
-                            borderRadius: BorderRadius.circular(10.0),
-                            border: Border.all(
-                              color: FlutterFlowTheme.of(context).sub1,
-                              width: 2.0,
-                            ),
-                          ),
-                          child: ClipRRect(
-                            borderRadius: BorderRadius.circular(8.0),
-                            child: Image.network(
-                              valueOrDefault<String>(
-                                uploadedImagesItem,
-                                'https://icons.veryicon.com/png/o/miscellaneous/1em/add-image.png',
-                              ),
-                              width: MediaQuery.sizeOf(context).width * 0.1,
-                              height: MediaQuery.sizeOf(context).height * 0.1,
-                              fit: BoxFit.cover,
-                            ),
-                          ),
-                        ),
-                      );
-                    }),
-                  ),
-                );
-              },
-            ),
-            Padding(
-              padding: EdgeInsetsDirectional.fromSTEB(16.0, 12.0, 16.0, 0.0),
-              child: Row(
+              padding: const EdgeInsetsDirectional.fromSTEB(20.0, 0.0, 0.0, 0.0),
+              child: Column(
                 mainAxisSize: MainAxisSize.max,
                 children: [
+                  Align(
+                    alignment: const AlignmentDirectional(-1.0, 0.0),
+                    child: Padding(
+                      padding:
+                          const EdgeInsetsDirectional.fromSTEB(0.0, 0.0, 0.0, 10.0),
+                      child: Text(
+                        'Description',
+                        style:
+                            FlutterFlowTheme.of(context).titleMedium.override(
+                                  fontFamily: FlutterFlowTheme.of(context)
+                                      .titleMediumFamily,
+                                  fontSize: 16.0,
+                                  useGoogleFonts: GoogleFonts.asMap()
+                                      .containsKey(FlutterFlowTheme.of(context)
+                                          .titleMediumFamily),
+                                ),
+                      ),
+                    ),
+                  ),
                   Padding(
                     padding:
-                        EdgeInsetsDirectional.fromSTEB(0.0, 0.0, 12.0, 0.0),
-                    child: FlutterFlowIconButton(
-                      borderColor: Colors.transparent,
-                      borderRadius: 30.0,
-                      borderWidth: 1.0,
-                      buttonSize: 60.0,
-                      icon: Icon(
-                        Icons.add_photo_alternate_outlined,
-                        color: FlutterFlowTheme.of(context).primary,
-                        size: 32.0,
+                        const EdgeInsetsDirectional.fromSTEB(0.0, 0.0, 20.0, 0.0),
+                    child: Container(
+                      width: MediaQuery.sizeOf(context).width * 1.0,
+                      height: MediaQuery.sizeOf(context).height * 0.17,
+                      decoration: BoxDecoration(
+                        borderRadius: BorderRadius.circular(7.0),
+                        border: Border.all(
+                          color: const Color(0xFFB8B6B6),
+                          width: 2.0,
+                        ),
                       ),
-                      showLoadingIndicator: true,
-                      onPressed: () async {
-                        // uploadPhoto
-                        final selectedMedia = await selectMedia(
-                          mediaSource: MediaSource.photoGallery,
-                          multiImage: true,
-                        );
-                        if (selectedMedia != null &&
-                            selectedMedia.every((m) =>
-                                validateFileFormat(m.storagePath, context))) {
-                          setState(() => _model.isDataUploading = true);
-                          var selectedUploadedFiles = <FFUploadedFile>[];
-
-                          var downloadUrls = <String>[];
-                          try {
-                            selectedUploadedFiles = selectedMedia
-                                .map((m) => FFUploadedFile(
-                                      name: m.storagePath.split('/').last,
-                                      bytes: m.bytes,
-                                      height: m.dimensions?.height,
-                                      width: m.dimensions?.width,
-                                      blurHash: m.blurHash,
-                                    ))
-                                .toList();
-
-                            downloadUrls = (await Future.wait(
-                              selectedMedia.map(
-                                (m) async =>
-                                    await uploadData(m.storagePath, m.bytes),
+                      alignment: const AlignmentDirectional(0.0, 0.0),
+                      child: TextFormField(
+                        controller: _model.textController3,
+                        focusNode: _model.textFieldFocusNode3,
+                        onChanged: (_) => EasyDebounce.debounce(
+                          '_model.textController3',
+                          const Duration(milliseconds: 2000),
+                          () => setState(() {}),
+                        ),
+                        textCapitalization: TextCapitalization.sentences,
+                        obscureText: false,
+                        decoration: InputDecoration(
+                          hintText:
+                              'Tell us about your item e.g. brand, \nmaterial, condition and size.\n\nInclude anything that you think your \nneighbours would like to know. ',
+                          hintStyle: FlutterFlowTheme.of(context)
+                              .labelLarge
+                              .override(
+                                fontFamily: FlutterFlowTheme.of(context)
+                                    .labelLargeFamily,
+                                color: const Color(0xFFA3A2A2),
+                                useGoogleFonts: GoogleFonts.asMap().containsKey(
+                                    FlutterFlowTheme.of(context)
+                                        .labelLargeFamily),
                               ),
-                            ))
-                                .where((u) => u != null)
-                                .map((u) => u!)
-                                .toList();
-                          } finally {
-                            _model.isDataUploading = false;
-                          }
-                          if (selectedUploadedFiles.length ==
-                                  selectedMedia.length &&
-                              downloadUrls.length == selectedMedia.length) {
-                            setState(() {
-                              _model.uploadedLocalFiles = selectedUploadedFiles;
-                              _model.uploadedFileUrls = downloadUrls;
-                            });
-                          } else {
-                            setState(() {});
-                            return;
-                          }
-                        }
-                      },
+                          enabledBorder: InputBorder.none,
+                          focusedBorder: InputBorder.none,
+                          errorBorder: InputBorder.none,
+                          focusedErrorBorder: InputBorder.none,
+                          contentPadding: const EdgeInsetsDirectional.fromSTEB(
+                              16.0, 0.0, 16.0, 0.0),
+                        ),
+                        style: FlutterFlowTheme.of(context).bodyLarge,
+                        textAlign: TextAlign.start,
+                        maxLines: 8,
+                        cursorColor: FlutterFlowTheme.of(context).backG1,
+                        validator: _model.textController3Validator
+                            .asValidator(context),
+                      ),
                     ),
                   ),
                 ],
               ),
             ),
             Padding(
-              padding: EdgeInsetsDirectional.fromSTEB(20.0, 180.0, 20.0, 20.0),
+              padding: const EdgeInsetsDirectional.fromSTEB(20.0, 0.0, 20.0, 8.0),
               child: InkWell(
                 splashColor: Colors.transparent,
                 focusColor: Colors.transparent,
                 hoverColor: Colors.transparent,
                 highlightColor: Colors.transparent,
                 onTap: () async {
-                  if ((_model.textController1.text != null &&
-                          _model.textController1.text != '') &&
-                      (_model.textController2.text != null &&
-                          _model.textController2.text != '') &&
-                      (_model.isDataUploading != null)) {
+                  if ((_model.textController1.text != '') &&
+                      (_model.textController1.text != '') &&
+                      (_model.textController3.text != '')) {
                     await PostRecord.collection.doc().set({
                       ...createPostRecordData(
                         category: _model.dropDownValue,
-                        title: _model.textController2.text,
-                        contents: _model.textController1.text,
+                        title: _model.textController1.text,
+                        contents: _model.textController3.text,
                         uploadTime: getCurrentTimestamp,
                         uploadUser: currentUserReference,
+                        price: 1250,
+                        location: 'Andijan',
+                        forSale: _model.choiceChipsValue == 'Free',
+                        negoitable: _model.checkboxValue,
+                        meetLocation: widget.location,
                       ),
-                      'uploadedImages': _model.uploadedFileUrls,
+                      ...mapToFirestore(
+                        {
+                          'uploadedImages': _model.uploadedFileUrls.isNotEmpty
+                              ? _model.uploadedFileUrls
+                              : FFAppState().isImage,
+                        },
+                      ),
                     });
                     await showModalBottomSheet(
                       isScrollControlled: true,
@@ -430,9 +799,9 @@ class _UploadPostWidgetState extends State<UploadPostWidget> {
                       builder: (context) {
                         return Padding(
                           padding: MediaQuery.viewInsetsOf(context),
-                          child: Container(
+                          child: SizedBox(
                             height: MediaQuery.sizeOf(context).height * 1.0,
-                            child: SucessPopUpWidget(),
+                            child: const SucessPopUpWidget(),
                           ),
                         );
                       },
@@ -446,9 +815,9 @@ class _UploadPostWidgetState extends State<UploadPostWidget> {
                       builder: (context) {
                         return Padding(
                           padding: MediaQuery.viewInsetsOf(context),
-                          child: Container(
+                          child: SizedBox(
                             height: MediaQuery.sizeOf(context).height * 1.0,
-                            child: FailedPopupWidget(),
+                            child: const FailedPopupWidget(),
                           ),
                         );
                       },
@@ -457,13 +826,13 @@ class _UploadPostWidgetState extends State<UploadPostWidget> {
                 },
                 child: Container(
                   width: MediaQuery.sizeOf(context).width * 1.0,
-                  height: 60.0,
+                  height: 50.0,
                   decoration: BoxDecoration(
                     color: FlutterFlowTheme.of(context).main1,
                     borderRadius: BorderRadius.circular(10.0),
                   ),
                   child: Align(
-                    alignment: AlignmentDirectional(0.00, 0.00),
+                    alignment: const AlignmentDirectional(0.0, 0.0),
                     child: Text(
                       'Post',
                       style: FlutterFlowTheme.of(context).titleLarge.override(
