@@ -2,6 +2,7 @@ import 'dart:convert';
 
 import 'package:flutter/material.dart';
 import 'package:from_css_color/from_css_color.dart';
+
 import '/flutter_flow/flutter_flow_util.dart';
 
 export 'package:collection/collection.dart' show ListEquality;
@@ -15,6 +16,40 @@ abstract class BaseStruct {
   String serialize() => json.encode(toSerializableMap());
 }
 
+dynamic deserializeStructParam<T>(
+  dynamic param,
+  ParamType paramType,
+  bool isList, {
+  required StructBuilder<T> structBuilder,
+}) {
+  if (param == null) {
+    return null;
+  } else if (isList) {
+    final paramValues;
+    try {
+      paramValues = param is Iterable ? param : json.decode(param);
+    } catch (e) {
+      return null;
+    }
+    if (paramValues is! Iterable) {
+      return null;
+    }
+    return paramValues
+        .map<T>((e) => deserializeStructParam<T>(e, paramType, false,
+            structBuilder: structBuilder))
+        .toList();
+  } else if (param is Map<String, dynamic>) {
+    return structBuilder(param);
+  } else {
+    return deserializeParam<T>(
+      param,
+      paramType,
+      isList,
+      structBuilder: structBuilder,
+    );
+  }
+}
+
 List<T>? getStructList<T>(
   dynamic value,
   StructBuilder<T> structBuilder,
@@ -22,8 +57,8 @@ List<T>? getStructList<T>(
     value is! List
         ? null
         : value
-            .where((e) => e is Map<String, dynamic>)
-            .map((e) => structBuilder(e as Map<String, dynamic>))
+            .whereType<Map<String, dynamic>>()
+            .map((e) => structBuilder(e))
             .toList();
 
 Color? getSchemaColor(dynamic value) => value is String
@@ -37,27 +72,6 @@ List<Color>? getColorsList(dynamic value) =>
 
 List<T>? getDataList<T>(dynamic value) =>
     value is! List ? null : value.map((e) => castToType<T>(e)!).toList();
-
-T? castToType<T>(dynamic value) {
-  if (value == null) {
-    return null;
-  }
-  switch (T) {
-    case double:
-      // Doubles may be stored as ints in some cases.
-      return value.toDouble() as T;
-    case int:
-      // Likewise, ints may be stored as doubles. If this is the case
-      // (i.e. no decimal value), return the value as an int.
-      if (value is num && value.toInt() == value) {
-        return value.toInt() as T;
-      }
-      break;
-    default:
-      break;
-  }
-  return value as T;
-}
 
 extension MapDataExtensions on Map<String, dynamic> {
   Map<String, dynamic> get withoutNulls => Map.fromEntries(
