@@ -5,18 +5,22 @@ import 'package:flutter/material.dart';
 
 class FlutterFlowDropDown<T> extends StatefulWidget {
   const FlutterFlowDropDown({
-    required this.controller,
+    super.key,
+    this.controller,
+    this.multiSelectController,
     this.hintText,
     this.searchHintText,
     required this.options,
     this.optionLabels,
     this.onChanged,
-    this.onChangedForMultiSelect,
+    this.onMultiSelectChanged,
     this.icon,
     this.width,
     this.height,
+    this.maxHeight,
     this.fillColor,
     this.searchHintTextStyle,
+    this.searchTextStyle,
     this.searchCursorColor,
     required this.textStyle,
     required this.elevation,
@@ -26,22 +30,37 @@ class FlutterFlowDropDown<T> extends StatefulWidget {
     required this.margin,
     this.hidesUnderline = false,
     this.disabled = false,
+    this.isOverButton = false,
+    this.menuOffset,
     this.isSearchable = false,
     this.isMultiSelect = false,
-  });
+  }) : assert(
+          isMultiSelect
+              ? (controller == null &&
+                  onChanged == null &&
+                  multiSelectController != null &&
+                  onMultiSelectChanged != null)
+              : (controller != null &&
+                  onChanged != null &&
+                  multiSelectController == null &&
+                  onMultiSelectChanged == null),
+        );
 
-  final FormFieldController<T> controller;
+  final FormFieldController<T?>? controller;
+  final FormFieldController<List<T>?>? multiSelectController;
   final String? hintText;
   final String? searchHintText;
   final List<T> options;
   final List<String>? optionLabels;
   final Function(T?)? onChanged;
-  final Function(List<T>?)? onChangedForMultiSelect;
+  final Function(List<T>?)? onMultiSelectChanged;
   final Widget? icon;
   final double? width;
   final double? height;
+  final double? maxHeight;
   final Color? fillColor;
   final TextStyle? searchHintTextStyle;
+  final TextStyle? searchTextStyle;
   final Color? searchCursorColor;
   final TextStyle textStyle;
   final double elevation;
@@ -51,6 +70,8 @@ class FlutterFlowDropDown<T> extends StatefulWidget {
   final EdgeInsetsGeometry margin;
   final bool hidesUnderline;
   final bool disabled;
+  final bool isOverButton;
+  final Offset? menuOffset;
   final bool isSearchable;
   final bool isMultiSelect;
 
@@ -59,111 +80,111 @@ class FlutterFlowDropDown<T> extends StatefulWidget {
 }
 
 class _FlutterFlowDropDownState<T> extends State<FlutterFlowDropDown<T>> {
-  final TextEditingController _textEditingController = TextEditingController();
+  bool get isMultiSelect => widget.isMultiSelect;
+  FormFieldController<T?> get controller => widget.controller!;
+  FormFieldController<List<T>?> get multiSelectController =>
+      widget.multiSelectController!;
 
-  void Function() get listener => widget.isMultiSelect
-      ? () {}
-      : () => widget.onChanged!(widget.controller.value);
+  T? get currentValue {
+    final value = isMultiSelect
+        ? multiSelectController.value?.firstOrNull
+        : controller.value;
+    return widget.options.contains(value) ? value : null;
+  }
+
+  Set<T> get currentValues {
+    if (!isMultiSelect || multiSelectController.value == null) {
+      return {};
+    }
+    return widget.options
+        .toSet()
+        .intersection(multiSelectController.value!.toSet());
+  }
+
+  Map<T, String> get optionLabels => Map.fromEntries(
+        widget.options.asMap().entries.map(
+              (option) => MapEntry(
+                option.value,
+                widget.optionLabels == null ||
+                        widget.optionLabels!.length < option.key + 1
+                    ? option.value.toString()
+                    : widget.optionLabels![option.key],
+              ),
+            ),
+      );
+
+  EdgeInsetsGeometry get horizontalMargin => widget.margin.clamp(
+        EdgeInsetsDirectional.zero,
+        const EdgeInsetsDirectional.symmetric(horizontal: double.infinity),
+      );
+
+  late void Function() _listener;
+  final TextEditingController _textEditingController = TextEditingController();
 
   @override
   void initState() {
-    widget.controller.addListener(listener);
     super.initState();
+    if (isMultiSelect) {
+      _listener =
+          () => widget.onMultiSelectChanged!(multiSelectController.value);
+      multiSelectController.addListener(_listener);
+    } else {
+      _listener = () => widget.onChanged!(controller.value);
+      controller.addListener(_listener);
+    }
   }
 
   @override
   void dispose() {
-    widget.controller.removeListener(listener);
+    if (isMultiSelect) {
+      multiSelectController.removeListener(_listener);
+    } else {
+      controller.removeListener(_listener);
+    }
     super.dispose();
   }
 
-  List<T> selectedItems = [];
-
   @override
   Widget build(BuildContext context) {
-    final optionToDisplayValue = Map.fromEntries(
-      widget.options.asMap().entries.map((option) => MapEntry(
-          option.value,
-          widget.optionLabels == null ||
-                  widget.optionLabels!.length < option.key + 1
-              ? option.value.toString()
-              : widget.optionLabels![option.key])),
-    );
-    final value = widget.options.contains(widget.controller.value)
-        ? widget.controller.value
-        : null;
-    final items = widget.options
-        .map(
-          (option) => DropdownMenuItem<T>(
-            value: option,
-            child: Text(
-              optionToDisplayValue[option] ?? '',
-              style: widget.textStyle,
-            ),
+    final dropdownWidget = _buildDropdownWidget();
+    return SizedBox(
+      width: widget.width,
+      height: widget.height,
+      child: DecoratedBox(
+        decoration: BoxDecoration(
+          borderRadius: BorderRadius.circular(widget.borderRadius),
+          border: Border.all(
+            color: widget.borderColor,
+            width: widget.borderWidth,
           ),
-        )
-        .toList();
-    final hintText = widget.hintText != null
-        ? Text(widget.hintText!, style: widget.textStyle)
-        : null;
-    void Function(T?)? onChanged = widget.disabled || widget.isMultiSelect
-        ? null
-        : (value) => widget.controller.value = value;
-    final dropdownWidget = widget.isMultiSelect
-        ? _buildMultiSelectDropdown(
-            hintText,
-            optionToDisplayValue,
-            widget.isSearchable,
-            widget.onChangedForMultiSelect!,
-            widget.disabled,
-          )
-        : widget.isSearchable
-            ? _buildSearchableDropdown(
-                value,
-                items,
-                onChanged,
-                hintText,
-                optionToDisplayValue,
-              )
-            : _buildNonSearchableDropdown(value, items, onChanged, hintText);
-    final childWidget = DecoratedBox(
-      decoration: BoxDecoration(
-        borderRadius: BorderRadius.circular(widget.borderRadius),
-        border: Border.all(
-          color: widget.borderColor,
-          width: widget.borderWidth,
+          color: widget.fillColor,
         ),
-        color: widget.fillColor,
-      ),
-      child: Padding(
-        padding: widget.margin,
-        child: widget.hidesUnderline
-            ? DropdownButtonHideUnderline(child: dropdownWidget)
-            : dropdownWidget,
+        child: Padding(
+          padding: _useDropdown2() ? EdgeInsets.zero : widget.margin,
+          child: widget.hidesUnderline
+              ? DropdownButtonHideUnderline(child: dropdownWidget)
+              : dropdownWidget,
+        ),
       ),
     );
-    if (widget.height != null || widget.width != null) {
-      return Container(
-        width: widget.width,
-        height: widget.height,
-        child: childWidget,
-      );
-    }
-    return childWidget;
   }
 
-  Widget _buildNonSearchableDropdown(
-    T? value,
-    List<DropdownMenuItem<T>>? items,
-    void Function(T?)? onChanged,
-    Text? hintText,
-  ) {
+  bool _useDropdown2() =>
+      widget.isMultiSelect ||
+      widget.isSearchable ||
+      !widget.isOverButton ||
+      widget.maxHeight != null;
+
+  Widget _buildDropdownWidget() =>
+      _useDropdown2() ? _buildDropdown() : _buildLegacyDropdown();
+
+  Widget _buildLegacyDropdown() {
     return DropdownButton<T>(
-      value: value,
-      hint: hintText,
-      items: items,
+      value: currentValue,
+      hint: _createHintText(),
+      items: _createMenuItems(),
       elevation: widget.elevation.toInt(),
-      onChanged: onChanged,
+      onChanged: widget.disabled ? null : (value) => controller.value = value,
       icon: widget.icon,
       isExpanded: true,
       dropdownColor: widget.fillColor,
@@ -171,120 +192,47 @@ class _FlutterFlowDropDownState<T> extends State<FlutterFlowDropDown<T>> {
     );
   }
 
-  Widget _buildSearchableDropdown(
-    T? value,
-    List<DropdownMenuItem<T>>? items,
-    void Function(T?)? onChanged,
-    Text? hintText,
-    Map<T, String> optionLabels,
-  ) {
-    final overlayColor = MaterialStateProperty.resolveWith<Color?>((states) =>
-        states.contains(MaterialState.focused) ? Colors.transparent : null);
-    final iconStyleData = widget.icon != null
-        ? IconStyleData(icon: widget.icon!)
-        : const IconStyleData();
-    return DropdownButton2<T>(
-      value: value,
-      hint: hintText,
-      items: items,
-      iconStyleData: iconStyleData,
-      buttonStyleData: ButtonStyleData(
-        elevation: widget.elevation.toInt(),
-        overlayColor: overlayColor,
-      ),
-      menuItemStyleData: MenuItemStyleData(overlayColor: overlayColor),
-      dropdownStyleData: DropdownStyleData(
-        elevation: widget.elevation.toInt(),
-        decoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(4),
-          color: widget.fillColor,
-        ),
-      ),
-      onChanged: onChanged,
-      isExpanded: true,
-      dropdownSearchData: DropdownSearchData<T>(
-        searchController: _textEditingController,
-        searchInnerWidgetHeight: 50,
-        searchInnerWidget: Container(
-          height: 50,
-          padding: const EdgeInsets.only(
-            top: 8,
-            bottom: 4,
-            right: 8,
-            left: 8,
-          ),
-          child: TextFormField(
-            expands: true,
-            maxLines: null,
-            controller: _textEditingController,
-            cursorColor: widget.searchCursorColor,
-            decoration: InputDecoration(
-              isDense: true,
-              contentPadding: const EdgeInsets.symmetric(
-                horizontal: 10,
-                vertical: 8,
-              ),
-              hintText: widget.searchHintText,
-              hintStyle: widget.searchHintTextStyle,
-              border: OutlineInputBorder(
-                borderRadius: BorderRadius.circular(8),
-              ),
-            ),
+  Text? _createHintText() => widget.hintText != null
+      ? Text(widget.hintText!, style: widget.textStyle)
+      : null;
+
+  List<DropdownMenuItem<T>> _createMenuItems() => widget.options
+      .map(
+        (option) => DropdownMenuItem<T>(
+          value: option,
+          child: Padding(
+            padding: _useDropdown2() ? horizontalMargin : EdgeInsets.zero,
+            child: Text(optionLabels[option] ?? '', style: widget.textStyle),
           ),
         ),
-        searchMatchFn: (item, searchValue) {
-          return (optionLabels[item.value] ?? '')
-              .toLowerCase()
-              .contains(searchValue.toLowerCase());
-        },
-      ),
+      )
+      .toList();
 
-      // This to clear the search value when you close the menu
-      onMenuStateChange: (isOpen) {
-        if (!isOpen) {
-          _textEditingController.clear();
-        }
-      },
-    );
-  }
-
-  Widget _buildMultiSelectDropdown(
-    Text? hintText,
-    Map<T, String> optionLabels,
-    bool isSearchable,
-    Function(List<T>?) onChangedForMultiSelect,
-    bool disabled,
-  ) {
-    final overlayColor = MaterialStateProperty.resolveWith<Color?>((states) =>
-        states.contains(MaterialState.focused) ? Colors.transparent : null);
-    final iconStyleData = widget.icon != null
-        ? IconStyleData(icon: widget.icon!)
-        : const IconStyleData();
-    return DropdownButton2<T>(
-      value: selectedItems.isEmpty ? null : selectedItems.last,
-      hint: hintText,
-      items: widget.options.map((item) {
-        return DropdownMenuItem(
+  List<DropdownMenuItem<T>> _createMultiselectMenuItems() => widget.options
+      .map(
+        (item) => DropdownMenuItem<T>(
           value: item,
           // Disable default onTap to avoid closing menu when selecting an item
           enabled: false,
           child: StatefulBuilder(
             builder: (context, menuSetState) {
-              final isSelected = selectedItems.contains(item);
+              final isSelected =
+                  multiSelectController.value?.contains(item) ?? false;
               return InkWell(
                 onTap: () {
+                  multiSelectController.value ??= [];
                   isSelected
-                      ? selectedItems.remove(item)
-                      : selectedItems.add(item);
-                  onChangedForMultiSelect(selectedItems);
-                  //This rebuilds the StatefulWidget to update the button's text
+                      ? multiSelectController.value!.remove(item)
+                      : multiSelectController.value!.add(item);
+                  multiSelectController.update();
+                  // This rebuilds the StatefulWidget to update the button's text.
                   setState(() {});
-                  //This rebuilds the dropdownMenu Widget to update the check mark
+                  // This rebuilds the dropdownMenu Widget to update the check mark.
                   menuSetState(() {});
                 },
                 child: Container(
                   height: double.infinity,
-                  padding: const EdgeInsets.symmetric(horizontal: 16.0),
+                  padding: horizontalMargin,
                   child: Row(
                     children: [
                       if (isSelected)
@@ -294,7 +242,7 @@ class _FlutterFlowDropDownState<T> extends State<FlutterFlowDropDown<T>> {
                       const SizedBox(width: 16),
                       Expanded(
                         child: Text(
-                          item as String,
+                          optionLabels[item]!,
                           style: widget.textStyle,
                         ),
                       ),
@@ -304,42 +252,63 @@ class _FlutterFlowDropDownState<T> extends State<FlutterFlowDropDown<T>> {
               );
             },
           ),
-        );
-      }).toList(),
+        ),
+      )
+      .toList();
+
+  Widget _buildDropdown() {
+    final overlayColor = MaterialStateProperty.resolveWith<Color?>((states) =>
+        states.contains(MaterialState.focused) ? Colors.transparent : null);
+    final iconStyleData = widget.icon != null
+        ? IconStyleData(icon: widget.icon!)
+        : const IconStyleData();
+    return DropdownButton2<T>(
+      value: currentValue,
+      hint: _createHintText(),
+      items: isMultiSelect ? _createMultiselectMenuItems() : _createMenuItems(),
       iconStyleData: iconStyleData,
       buttonStyleData: ButtonStyleData(
-        elevation: widget.elevation!.toInt(),
+        elevation: widget.elevation.toInt(),
         overlayColor: overlayColor,
+        padding: widget.margin,
       ),
-      menuItemStyleData: MenuItemStyleData(overlayColor: overlayColor),
+      menuItemStyleData: MenuItemStyleData(
+        overlayColor: overlayColor,
+        padding: EdgeInsets.zero,
+      ),
       dropdownStyleData: DropdownStyleData(
-        elevation: widget.elevation!.toInt(),
+        elevation: widget.elevation.toInt(),
         decoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(4),
+          borderRadius: BorderRadius.circular(4.0),
           color: widget.fillColor,
         ),
+        isOverButton: widget.isOverButton,
+        offset: widget.menuOffset ?? Offset.zero,
+        maxHeight: widget.maxHeight,
+        padding: EdgeInsets.zero,
       ),
-      // onChanged is handled by the onChangedForMultiSelect function
-      onChanged: disabled ? null : (val) {},
+      onChanged: widget.disabled
+          ? null
+          : (isMultiSelect ? (_) {} : (val) => widget.controller!.value = val),
       isExpanded: true,
-      selectedItemBuilder: (context) {
-        return widget.options.map(
-          (item) {
-            return Container(
-              alignment: AlignmentDirectional.center,
+      selectedItemBuilder: (context) => widget.options
+          .map(
+            (item) => Align(
+              alignment: AlignmentDirectional.centerStart,
               child: Text(
-                selectedItems.join(', '),
-                style: const TextStyle(
-                  fontSize: 14,
-                  overflow: TextOverflow.ellipsis,
-                ),
+                isMultiSelect
+                    ? currentValues
+                        .where((v) => optionLabels.containsKey(v))
+                        .map((v) => optionLabels[v])
+                        .join(', ')
+                    : optionLabels[item]!,
+                style: widget.textStyle,
                 maxLines: 1,
               ),
-            );
-          },
-        ).toList();
-      },
-      dropdownSearchData: isSearchable
+            ),
+          )
+          .toList(),
+      dropdownSearchData: widget.isSearchable
           ? DropdownSearchData<T>(
               searchController: _textEditingController,
               searchInnerWidgetHeight: 50,
@@ -356,6 +325,7 @@ class _FlutterFlowDropDownState<T> extends State<FlutterFlowDropDown<T>> {
                   maxLines: null,
                   controller: _textEditingController,
                   cursorColor: widget.searchCursorColor,
+                  style: widget.searchTextStyle,
                   decoration: InputDecoration(
                     isDense: true,
                     contentPadding: const EdgeInsets.symmetric(
@@ -377,8 +347,8 @@ class _FlutterFlowDropDownState<T> extends State<FlutterFlowDropDown<T>> {
               },
             )
           : null,
-      // This to clear the search value when you close the menu
-      onMenuStateChange: isSearchable
+      // This is to clear the search value when you close the menu
+      onMenuStateChange: widget.isSearchable
           ? (isOpen) {
               if (!isOpen) {
                 _textEditingController.clear();
