@@ -1,11 +1,9 @@
 import 'dart:convert';
 
 import 'package:flutter/material.dart';
-import 'package:from_css_color/from_css_color.dart';
 
 import '/backend/backend.dart';
 
-import '../../flutter_flow/lat_lng.dart';
 import '../../flutter_flow/place.dart';
 import '../../flutter_flow/uploaded_file.dart';
 
@@ -88,6 +86,9 @@ String? serializeParam(
       case ParamType.Document:
         final reference = (param as FirestoreRecord).reference;
         return _serializeDocumentReference(reference);
+
+      case ParamType.DataStruct:
+        return param is BaseStruct ? param.serialize() : null;
 
       default:
         return null;
@@ -177,6 +178,7 @@ enum ParamType {
   JSON,
   Document,
   DocumentReference,
+  DataStruct,
 }
 
 dynamic deserializeParam<T>(
@@ -184,6 +186,7 @@ dynamic deserializeParam<T>(
   ParamType paramType,
   bool isList, {
   List<String>? collectionNamePath,
+  StructBuilder<T>? structBuilder,
 }) {
   try {
     if (param == null) {
@@ -195,10 +198,15 @@ dynamic deserializeParam<T>(
         return null;
       }
       return paramValues
-          .where((p) => p is String)
-          .map((p) => p as String)
-          .map((p) => deserializeParam<T>(p, paramType, false,
-              collectionNamePath: collectionNamePath))
+          .whereType<String>()
+          .map((p) => p)
+          .map((p) => deserializeParam<T>(
+                p,
+                paramType,
+                false,
+                collectionNamePath: collectionNamePath,
+                structBuilder: structBuilder,
+              ))
           .where((p) => p != null)
           .map((p) => p! as T)
           .toList();
@@ -232,6 +240,10 @@ dynamic deserializeParam<T>(
       case ParamType.DocumentReference:
         return _deserializeDocumentReference(param, collectionNamePath ?? []);
 
+      case ParamType.DataStruct:
+        final data = json.decode(param) as Map<String, dynamic>? ?? {};
+        return structBuilder != null ? structBuilder(data) : null;
+
       default:
         return null;
     }
@@ -258,7 +270,7 @@ Future<List<T>> Function(String) getDocList<T>(
     List<String> docIds = [];
     try {
       final ids = json.decode(idsList) as Iterable;
-      docIds = ids.where((d) => d is String).map((d) => d as String).toList();
+      docIds = ids.whereType<String>().map((d) => d).toList();
     } catch (_) {}
     return Future.wait(
       docIds.map(
