@@ -4,12 +4,24 @@ export '/pages/upload_post/upload_post_widget.dart' show UploadPostWidget;
 export '/all_chats/all_chats_widget.dart' show AllChatsWidget;
 export '/chat_page/chat_page_widget.dart' show ChatPageWidget;
 export '/pages/home_page/home_page_widget.dart' show HomePageWidget;
-export '/pages/profile/profile_widget.dart' show ProfileWidget;
 export '/pages/post_detail_page/post_detail_page_widget.dart'
     show PostDetailPageWidget;
-export '/pages/search_page/search_page_widget.dart' show SearchPageWidget;
 export '/pages/community_page/community_page_widget.dart'
     show CommunityPageWidget;
 export '/pages/c_t_detail_page/c_t_detail_page_widget.dart'
     show CTDetailPageWidget;
-export '/pages/c_upload_page/c_upload_page_widget.dart' show CUploadPageWidget;
+export '/edit_profile/edit_profile_widget.dart' show EditProfileWidget;
+export '/pages/account/account_widget.dart' show AccountWidget;
+export '/nearby/nearby_widget.dart' show NearbyWidget;
+export '/f_a_qspage/f_a_qspage_widget.dart' show FAQspageWidget;
+export '/real_estate/real_estate_widget.dart' show RealEstateWidget;
+export '/show_image/show_image_widget.dart' show ShowImageWidget;
+export '/village_post/village_post_widget.dart' show VillagePostWidget;
+export '/mappage/mappage_widget.dart' show MappageWidget;
+export '/mappage_copy/mappage_copy_widget.dart' show MappageCopyWidget;
+export '/sample/sample_widget.dart' show SampleWidget;
+export '/show_image_copy/show_image_copy_widget.dart' show ShowImageCopyWidget;
+export '/upload_ad/upload_ad_widget.dart' show UploadAdWidget;
+export '/ad_detail_page/ad_detail_page_widget.dart' show AdDetailPageWidget;
+export '/village_post_copy/village_post_copy_widget.dart'
+    show VillagePostCopyWidget;
