@@ -1,14 +1,11 @@
 import '/auth/firebase_auth/auth_util.dart';
 import '/backend/backend.dart';
-import '/flutter_flow/flutter_flow_animations.dart';
 import '/flutter_flow/flutter_flow_theme.dart';
 import '/flutter_flow/flutter_flow_toggle_icon.dart';
 import '/flutter_flow/flutter_flow_util.dart';
-import '/flutter_flow/flutter_flow_widgets.dart';
-import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/scheduler.dart';
-import 'package:flutter_animate/flutter_animate.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:google_fonts/google_fonts.dart';
@@ -18,9 +15,9 @@ export 'c_t_detail_page_model.dart';
 
 class CTDetailPageWidget extends StatefulWidget {
   const CTDetailPageWidget({
-    Key? key,
+    super.key,
     required this.communityTalkRef,
-  }) : super(key: key);
+  });
 
   final DocumentReference? communityTalkRef;
 
@@ -28,90 +25,10 @@ class CTDetailPageWidget extends StatefulWidget {
   _CTDetailPageWidgetState createState() => _CTDetailPageWidgetState();
 }
 
-class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
-    with TickerProviderStateMixin {
+class _CTDetailPageWidgetState extends State<CTDetailPageWidget> {
   late CTDetailPageModel _model;
 
   final scaffoldKey = GlobalKey<ScaffoldState>();
-
-  final animationsMap = {
-    'rowOnPageLoadAnimation': AnimationInfo(
-      trigger: AnimationTrigger.onPageLoad,
-      effects: [
-        FadeEffect(
-          curve: Curves.easeInOut,
-          delay: 0.ms,
-          duration: 600.ms,
-          begin: 0.0,
-          end: 1.0,
-        ),
-        MoveEffect(
-          curve: Curves.easeInOut,
-          delay: 0.ms,
-          duration: 600.ms,
-          begin: Offset(0.0, 20.0),
-          end: Offset(0.0, 0.0),
-        ),
-      ],
-    ),
-    'columnOnPageLoadAnimation1': AnimationInfo(
-      trigger: AnimationTrigger.onPageLoad,
-      effects: [
-        FadeEffect(
-          curve: Curves.easeInOut,
-          delay: 400.ms,
-          duration: 600.ms,
-          begin: 0.0,
-          end: 1.0,
-        ),
-        MoveEffect(
-          curve: Curves.easeInOut,
-          delay: 400.ms,
-          duration: 600.ms,
-          begin: Offset(0.0, 20.0),
-          end: Offset(0.0, 0.0),
-        ),
-      ],
-    ),
-    'columnOnPageLoadAnimation2': AnimationInfo(
-      trigger: AnimationTrigger.onPageLoad,
-      effects: [
-        FadeEffect(
-          curve: Curves.easeInOut,
-          delay: 600.ms,
-          duration: 600.ms,
-          begin: 0.0,
-          end: 1.0,
-        ),
-        MoveEffect(
-          curve: Curves.easeInOut,
-          delay: 600.ms,
-          duration: 600.ms,
-          begin: Offset(0.0, 20.0),
-          end: Offset(0.0, 0.0),
-        ),
-      ],
-    ),
-    'columnOnPageLoadAnimation3': AnimationInfo(
-      trigger: AnimationTrigger.onPageLoad,
-      effects: [
-        FadeEffect(
-          curve: Curves.easeInOut,
-          delay: 600.ms,
-          duration: 600.ms,
-          begin: 0.0,
-          end: 1.0,
-        ),
-        MoveEffect(
-          curve: Curves.easeInOut,
-          delay: 600.ms,
-          duration: 600.ms,
-          begin: Offset(0.0, 20.0),
-          end: Offset(0.0, 0.0),
-        ),
-      ],
-    ),
-  };
 
   @override
   void initState() {
@@ -119,6 +36,7 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
     _model = createModel(context, () => CTDetailPageModel());
 
     _model.textController ??= TextEditingController();
+    _model.textFieldFocusNode ??= FocusNode();
 
     WidgetsBinding.instance.addPostFrameCallback((_) => setState(() {}));
   }
@@ -132,6 +50,15 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
 
   @override
   Widget build(BuildContext context) {
+    if (isiOS) {
+      SystemChrome.setSystemUIOverlayStyle(
+        SystemUiOverlayStyle(
+          statusBarBrightness: Theme.of(context).brightness,
+          systemStatusBarContrastEnforced: true,
+        ),
+      );
+    }
+
     context.watch<FFAppState>();
 
     return StreamBuilder<CommunityTalkRecord>(
@@ -140,7 +67,6 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
         // Customize what your widget looks like when it's loading.
         if (!snapshot.hasData) {
           return Scaffold(
-            backgroundColor: FlutterFlowTheme.of(context).primaryBackground,
             body: Center(
               child: SizedBox(
                 width: 50.0,
@@ -155,10 +81,11 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
         }
         final cTDetailPageCommunityTalkRecord = snapshot.data!;
         return GestureDetector(
-          onTap: () => FocusScope.of(context).requestFocus(_model.unfocusNode),
+          onTap: () => _model.unfocusNode.canRequestFocus
+              ? FocusScope.of(context).requestFocus(_model.unfocusNode)
+              : FocusScope.of(context).unfocus(),
           child: Scaffold(
             key: scaffoldKey,
-            backgroundColor: FlutterFlowTheme.of(context).primaryBackground,
             body: SafeArea(
               top: true,
               child: Column(
@@ -175,7 +102,7 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                         child: Column(
                           mainAxisSize: MainAxisSize.max,
                           children: [
-                            Container(
+                            SizedBox(
                               height: 260.0,
                               child: Stack(
                                 children: [
@@ -189,23 +116,40 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                     ),
                                     child: Stack(
                                       children: [
-                                        Image.network(
-                                          cTDetailPageCommunityTalkRecord
-                                              .uploadImage,
-                                          width:
-                                              MediaQuery.sizeOf(context).width *
-                                                  1.0,
-                                          height: MediaQuery.sizeOf(context)
-                                                  .height *
-                                              1.0,
-                                          fit: BoxFit.cover,
+                                        InkWell(
+                                          splashColor: Colors.transparent,
+                                          focusColor: Colors.transparent,
+                                          hoverColor: Colors.transparent,
+                                          highlightColor: Colors.transparent,
+                                          onTap: () async {
+                                            context.pushNamed(
+                                              'showImageCopy',
+                                              queryParameters: {
+                                                'postDocument': serializeParam(
+                                                  widget.communityTalkRef,
+                                                  ParamType.DocumentReference,
+                                                ),
+                                              }.withoutNulls,
+                                            );
+                                          },
+                                          child: Image.network(
+                                            cTDetailPageCommunityTalkRecord
+                                                .uploadedImages.first,
+                                            width: MediaQuery.sizeOf(context)
+                                                    .width *
+                                                1.0,
+                                            height: MediaQuery.sizeOf(context)
+                                                    .height *
+                                                1.0,
+                                            fit: BoxFit.cover,
+                                          ),
                                         ),
                                         Align(
                                           alignment:
-                                              AlignmentDirectional(1.00, -1.00),
+                                              const AlignmentDirectional(1.0, -1.0),
                                           child: Padding(
                                             padding:
-                                                EdgeInsetsDirectional.fromSTEB(
+                                                const EdgeInsetsDirectional.fromSTEB(
                                                     0.0, 15.0, 15.0, 0.0),
                                             child: ToggleIcon(
                                               onPressed: () async {
@@ -218,14 +162,14 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                                 Icons.favorite_outlined,
                                                 color:
                                                     FlutterFlowTheme.of(context)
-                                                        .alternate,
+                                                        .main1,
                                                 size: 25.0,
                                               ),
                                               offIcon: Icon(
                                                 Icons.favorite_border,
                                                 color:
                                                     FlutterFlowTheme.of(context)
-                                                        .alternate,
+                                                        .main1,
                                                 size: 25.0,
                                               ),
                                             ),
@@ -233,7 +177,7 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                         ),
                                         Padding(
                                           padding:
-                                              EdgeInsetsDirectional.fromSTEB(
+                                              const EdgeInsetsDirectional.fromSTEB(
                                                   15.0, 15.0, 0.0, 0.0),
                                           child: InkWell(
                                             splashColor: Colors.transparent,
@@ -245,7 +189,9 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                             },
                                             child: Icon(
                                               Icons.chevron_left_outlined,
-                                              color: Colors.white,
+                                              color:
+                                                  FlutterFlowTheme.of(context)
+                                                      .main1,
                                               size: 30.0,
                                             ),
                                           ),
@@ -254,12 +200,12 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                     ),
                                   ),
                                   Align(
-                                    alignment: AlignmentDirectional(0.00, 1.00),
+                                    alignment: const AlignmentDirectional(0.0, 1.0),
                                     child: Container(
                                       width: MediaQuery.sizeOf(context).width *
                                           1.0,
                                       height: 30.0,
-                                      decoration: BoxDecoration(
+                                      decoration: const BoxDecoration(
                                         color: Colors.white,
                                         borderRadius: BorderRadius.only(
                                           bottomLeft: Radius.circular(0.0),
@@ -273,7 +219,7 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                         children: [
                                           Padding(
                                             padding:
-                                                EdgeInsetsDirectional.fromSTEB(
+                                                const EdgeInsetsDirectional.fromSTEB(
                                                     0.0, 0.5, 0.0, 0.0),
                                             child: Container(
                                               width: 70.0,
@@ -299,7 +245,7 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                     .secondaryBackground,
                               ),
                               child: Padding(
-                                padding: EdgeInsetsDirectional.fromSTEB(
+                                padding: const EdgeInsetsDirectional.fromSTEB(
                                     15.0, 0.0, 15.0, 0.0),
                                 child: Column(
                                   mainAxisSize: MainAxisSize.max,
@@ -372,10 +318,10 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                                                 context)
                                                             .width *
                                                         1.0,
-                                                    decoration: BoxDecoration(),
+                                                    decoration: const BoxDecoration(),
                                                     child: Padding(
                                                       padding:
-                                                          EdgeInsetsDirectional
+                                                          const EdgeInsetsDirectional
                                                               .fromSTEB(
                                                                   0.0,
                                                                   5.0,
@@ -391,20 +337,24 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                                             clipBehavior:
                                                                 Clip.antiAlias,
                                                             decoration:
-                                                                BoxDecoration(
+                                                                const BoxDecoration(
                                                               shape: BoxShape
                                                                   .circle,
                                                             ),
                                                             child:
                                                                 Image.network(
-                                                              containerUserRecord
-                                                                  .photoUrl,
+                                                              valueOrDefault<
+                                                                  String>(
+                                                                containerUserRecord
+                                                                    .photoUrl,
+                                                                'https://images.pexels.com/photos/771742/pexels-photo-771742.jpeg?auto=compress&cs=tinysrgb&dpr=1&w=500',
+                                                              ),
                                                               fit: BoxFit.cover,
                                                             ),
                                                           ),
                                                           Padding(
                                                             padding:
-                                                                EdgeInsetsDirectional
+                                                                const EdgeInsetsDirectional
                                                                     .fromSTEB(
                                                                         5.0,
                                                                         0.0,
@@ -412,24 +362,10 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                                                         0.0),
                                                             child: Text(
                                                               containerUserRecord
-                                                                  .displayName,
+                                                                  .email,
                                                               style: FlutterFlowTheme
                                                                       .of(context)
-                                                                  .bodyMedium
-                                                                  .override(
-                                                                    fontFamily:
-                                                                        FlutterFlowTheme.of(context)
-                                                                            .bodyMediumFamily,
-                                                                    fontSize:
-                                                                        12.0,
-                                                                    fontWeight:
-                                                                        FontWeight
-                                                                            .w300,
-                                                                    useGoogleFonts: GoogleFonts
-                                                                            .asMap()
-                                                                        .containsKey(
-                                                                            FlutterFlowTheme.of(context).bodyMediumFamily),
-                                                                  ),
+                                                                  .labelLarge,
                                                             ),
                                                           ),
                                                         ],
@@ -438,41 +374,46 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                                   );
                                                 },
                                               ),
-                                              Padding(
-                                                padding: EdgeInsetsDirectional
-                                                    .fromSTEB(
-                                                        0.0, 12.0, 0.0, 0.0),
-                                                child: Text(
-                                                  cTDetailPageCommunityTalkRecord
-                                                      .content,
-                                                  style: FlutterFlowTheme.of(
-                                                          context)
-                                                      .bodyMedium
-                                                      .override(
-                                                        fontFamily:
-                                                            FlutterFlowTheme.of(
-                                                                    context)
-                                                                .bodyMediumFamily,
-                                                        fontSize: 20.0,
-                                                        fontWeight:
-                                                            FontWeight.w500,
-                                                        useGoogleFonts: GoogleFonts
-                                                                .asMap()
-                                                            .containsKey(
-                                                                FlutterFlowTheme.of(
-                                                                        context)
-                                                                    .bodyMediumFamily),
-                                                      ),
+                                              Container(
+                                                width:
+                                                    MediaQuery.sizeOf(context)
+                                                            .width *
+                                                        0.95,
+                                                decoration: const BoxDecoration(),
+                                                child: Padding(
+                                                  padding: const EdgeInsetsDirectional
+                                                      .fromSTEB(
+                                                          0.0, 12.0, 0.0, 0.0),
+                                                  child: AutoSizeText(
+                                                    cTDetailPageCommunityTalkRecord
+                                                        .content,
+                                                    style: FlutterFlowTheme.of(
+                                                            context)
+                                                        .labelMedium
+                                                        .override(
+                                                          fontFamily:
+                                                              FlutterFlowTheme.of(
+                                                                      context)
+                                                                  .labelMediumFamily,
+                                                          fontWeight:
+                                                              FontWeight.normal,
+                                                          useGoogleFonts: GoogleFonts
+                                                                  .asMap()
+                                                              .containsKey(
+                                                                  FlutterFlowTheme.of(
+                                                                          context)
+                                                                      .labelMediumFamily),
+                                                        ),
+                                                  ),
                                                 ),
                                               ),
                                             ],
                                           ),
                                         ),
                                       ],
-                                    ).animateOnPageLoad(animationsMap[
-                                        'rowOnPageLoadAnimation']!),
+                                    ),
                                     Padding(
-                                      padding: EdgeInsetsDirectional.fromSTEB(
+                                      padding: const EdgeInsetsDirectional.fromSTEB(
                                           0.0, 15.0, 0.0, 0.0),
                                       child: Column(
                                         mainAxisSize: MainAxisSize.max,
@@ -491,7 +432,7 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                                       FlutterFlowTheme.of(
                                                               context)
                                                           .bodyMediumFamily,
-                                                  color: Color(0xFF9D9B9B),
+                                                  color: const Color(0xFF9D9B9B),
                                                   fontWeight: FontWeight.normal,
                                                   useGoogleFonts: GoogleFonts
                                                           .asMap()
@@ -502,11 +443,10 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                                 ),
                                           ),
                                         ],
-                                      ).animateOnPageLoad(animationsMap[
-                                          'columnOnPageLoadAnimation1']!),
+                                      ),
                                     ),
                                     Padding(
-                                      padding: EdgeInsetsDirectional.fromSTEB(
+                                      padding: const EdgeInsetsDirectional.fromSTEB(
                                           0.0, 20.0, 0.0, 0.0),
                                       child: Container(
                                         width:
@@ -520,10 +460,10 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                           mainAxisSize: MainAxisSize.max,
                                           children: [
                                             Align(
-                                              alignment: AlignmentDirectional(
-                                                  -1.00, 0.00),
+                                              alignment: const AlignmentDirectional(
+                                                  -1.0, 0.0),
                                               child: Padding(
-                                                padding: EdgeInsetsDirectional
+                                                padding: const EdgeInsetsDirectional
                                                     .fromSTEB(
                                                         0.0, 0.0, 0.0, 15.0),
                                                 child: Text(
@@ -553,7 +493,13 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                               stream: queryCommentRecord(
                                                 queryBuilder: (commentRecord) =>
                                                     commentRecord
-                                                        .orderBy('commentTime'),
+                                                        .where(
+                                                          'communitTalkRef',
+                                                          isEqualTo: widget
+                                                              .communityTalkRef,
+                                                        )
+                                                        .orderBy('commentTime',
+                                                            descending: true),
                                               ),
                                               builder: (context, snapshot) {
                                                 // Customize what your widget looks like when it's loading.
@@ -591,7 +537,7 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                                               .width *
                                                           1.0,
                                                       decoration:
-                                                          BoxDecoration(),
+                                                          const BoxDecoration(),
                                                       child: StreamBuilder<
                                                           UserRecord>(
                                                         stream: UserRecord
@@ -662,7 +608,7 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                                                   clipBehavior:
                                                                       Clip.antiAlias,
                                                                   decoration:
-                                                                      BoxDecoration(
+                                                                      const BoxDecoration(
                                                                     shape: BoxShape
                                                                         .circle,
                                                                   ),
@@ -687,7 +633,7 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                                                   ),
                                                                   child:
                                                                       Padding(
-                                                                    padding: EdgeInsetsDirectional
+                                                                    padding: const EdgeInsetsDirectional
                                                                         .fromSTEB(
                                                                             15.0,
                                                                             0.0,
@@ -709,7 +655,7 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                                                               MainAxisAlignment.spaceBetween,
                                                                           children: [
                                                                             Text(
-                                                                              rowUserRecord.displayName,
+                                                                              rowUserRecord.email,
                                                                               style: FlutterFlowTheme.of(context).bodyMedium.override(
                                                                                     fontFamily: FlutterFlowTheme.of(context).bodyMediumFamily,
                                                                                     fontWeight: FontWeight.w600,
@@ -735,7 +681,7 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                                                           ],
                                                                         ),
                                                                         Padding(
-                                                                          padding: EdgeInsetsDirectional.fromSTEB(
+                                                                          padding: const EdgeInsetsDirectional.fromSTEB(
                                                                               0.0,
                                                                               7.0,
                                                                               0.0,
@@ -748,12 +694,12 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                                                           ),
                                                                         ),
                                                                         Align(
-                                                                          alignment: AlignmentDirectional(
-                                                                              1.00,
-                                                                              1.00),
+                                                                          alignment: const AlignmentDirectional(
+                                                                              1.0,
+                                                                              1.0),
                                                                           child:
                                                                               Padding(
-                                                                            padding: EdgeInsetsDirectional.fromSTEB(
+                                                                            padding: const EdgeInsetsDirectional.fromSTEB(
                                                                                 0.0,
                                                                                 5.0,
                                                                                 0.0,
@@ -763,7 +709,7 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                                                               dateTimeFormat('jm', columnCommentRecord.commentTime!),
                                                                               style: FlutterFlowTheme.of(context).bodyMedium.override(
                                                                                     fontFamily: FlutterFlowTheme.of(context).bodyMediumFamily,
-                                                                                    color: Color(0xFF959595),
+                                                                                    color: const Color(0xFF959595),
                                                                                     useGoogleFonts: GoogleFonts.asMap().containsKey(FlutterFlowTheme.of(context).bodyMediumFamily),
                                                                                   ),
                                                                             ),
@@ -780,8 +726,7 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                                       ),
                                                     );
                                                   }),
-                                                ).animateOnPageLoad(animationsMap[
-                                                    'columnOnPageLoadAnimation2']!);
+                                                );
                                               },
                                             ),
                                           ],
@@ -796,7 +741,7 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                             .secondaryBackground,
                                       ),
                                       child: Padding(
-                                        padding: EdgeInsetsDirectional.fromSTEB(
+                                        padding: const EdgeInsetsDirectional.fromSTEB(
                                             0.0, 30.0, 0.0, 20.0),
                                         child: Column(
                                           mainAxisSize: MainAxisSize.max,
@@ -813,7 +758,7 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                                                     context)
                                                                 .bodyMediumFamily,
                                                         color:
-                                                            Color(0xFF868686),
+                                                            const Color(0xFF868686),
                                                         useGoogleFonts: GoogleFonts
                                                                 .asMap()
                                                             .containsKey(
@@ -823,7 +768,7 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                                       ),
                                             ),
                                             Padding(
-                                              padding: EdgeInsetsDirectional
+                                              padding: const EdgeInsetsDirectional
                                                   .fromSTEB(
                                                       0.0, 10.0, 0.0, 0.0),
                                               child: Text(
@@ -851,8 +796,7 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                                               ),
                                             ),
                                           ],
-                                        ).animateOnPageLoad(animationsMap[
-                                            'columnOnPageLoadAnimation3']!),
+                                        ),
                                       ),
                                     ),
                                   ],
@@ -870,106 +814,96 @@ class _CTDetailPageWidgetState extends State<CTDetailPageWidget>
                     child: Container(
                       width: MediaQuery.sizeOf(context).width * 1.0,
                       height: 50.0,
-                      decoration: BoxDecoration(
+                      decoration: const BoxDecoration(
                         color: Colors.white,
                       ),
                       child: Padding(
-                        padding: EdgeInsetsDirectional.fromSTEB(
+                        padding: const EdgeInsetsDirectional.fromSTEB(
                             15.0, 0.0, 15.0, 0.0),
                         child: Row(
                           mainAxisSize: MainAxisSize.max,
                           children: [
                             Expanded(
                               child: Padding(
-                                padding: EdgeInsetsDirectional.fromSTEB(
-                                    8.0, 8.0, 8.0, 8.0),
+                                padding: const EdgeInsetsDirectional.fromSTEB(
+                                    8.0, 0.0, 8.0, 8.0),
                                 child: TextFormField(
                                   controller: _model.textController,
+                                  focusNode: _model.textFieldFocusNode,
                                   obscureText: false,
                                   decoration: InputDecoration(
-                                    labelText: '댓글 입력',
+                                    labelText: 'Leave a comment ',
                                     labelStyle: FlutterFlowTheme.of(context)
-                                        .labelMedium
-                                        .override(
-                                          fontFamily:
-                                              FlutterFlowTheme.of(context)
-                                                  .labelMediumFamily,
-                                          fontSize: 10.0,
-                                          useGoogleFonts: GoogleFonts.asMap()
-                                              .containsKey(
-                                                  FlutterFlowTheme.of(context)
-                                                      .labelMediumFamily),
-                                        ),
+                                        .labelMedium,
                                     hintStyle: FlutterFlowTheme.of(context)
                                         .labelMedium,
                                     enabledBorder: OutlineInputBorder(
                                       borderSide: BorderSide(
-                                        color: FlutterFlowTheme.of(context)
-                                            .warning,
-                                        width: 1.0,
+                                        color:
+                                            FlutterFlowTheme.of(context).main1,
+                                        width: 2.0,
                                       ),
-                                      borderRadius: BorderRadius.circular(8.0),
+                                      borderRadius: BorderRadius.circular(10.0),
                                     ),
                                     focusedBorder: OutlineInputBorder(
                                       borderSide: BorderSide(
                                         color:
-                                            FlutterFlowTheme.of(context).main3,
-                                        width: 1.0,
+                                            FlutterFlowTheme.of(context).main1,
+                                        width: 2.0,
                                       ),
-                                      borderRadius: BorderRadius.circular(8.0),
+                                      borderRadius: BorderRadius.circular(10.0),
                                     ),
                                     errorBorder: OutlineInputBorder(
                                       borderSide: BorderSide(
                                         color:
                                             FlutterFlowTheme.of(context).error,
-                                        width: 1.0,
+                                        width: 2.0,
                                       ),
-                                      borderRadius: BorderRadius.circular(8.0),
+                                      borderRadius: BorderRadius.circular(10.0),
                                     ),
                                     focusedErrorBorder: OutlineInputBorder(
                                       borderSide: BorderSide(
                                         color:
                                             FlutterFlowTheme.of(context).error,
-                                        width: 1.0,
+                                        width: 2.0,
                                       ),
-                                      borderRadius: BorderRadius.circular(8.0),
+                                      borderRadius: BorderRadius.circular(10.0),
                                     ),
                                   ),
                                   style:
                                       FlutterFlowTheme.of(context).bodyMedium,
-                                  maxLines: 2,
-                                  cursorColor:
-                                      FlutterFlowTheme.of(context).sub1,
                                   validator: _model.textControllerValidator
                                       .asValidator(context),
                                 ),
                               ),
                             ),
-                            InkWell(
-                              splashColor: Colors.transparent,
-                              focusColor: Colors.transparent,
-                              hoverColor: Colors.transparent,
-                              highlightColor: Colors.transparent,
-                              onTap: () async {
-                                await CommentRecord.collection
-                                    .doc()
-                                    .set(createCommentRecordData(
-                                      userRef: currentUserReference,
-                                      comment: valueOrDefault<String>(
-                                        _model.textController.text,
-                                        '-',
-                                      ),
-                                      commentTime: getCurrentTimestamp,
-                                      communitTalkRef: widget.communityTalkRef,
-                                    ));
-                                setState(() {
-                                  _model.textController?.clear();
-                                });
-                              },
-                              child: Icon(
-                                Icons.send_rounded,
-                                color: FlutterFlowTheme.of(context).sub1,
-                                size: 18.0,
+                            Padding(
+                              padding: const EdgeInsetsDirectional.fromSTEB(
+                                  0.0, 0.0, 0.0, 8.0),
+                              child: InkWell(
+                                splashColor: Colors.transparent,
+                                focusColor: Colors.transparent,
+                                hoverColor: Colors.transparent,
+                                highlightColor: Colors.transparent,
+                                onTap: () async {
+                                  await CommentRecord.collection
+                                      .doc()
+                                      .set(createCommentRecordData(
+                                        userRef: currentUserReference,
+                                        comment: _model.textController.text,
+                                        commentTime: getCurrentTimestamp,
+                                        communitTalkRef:
+                                            widget.communityTalkRef,
+                                      ));
+                                  setState(() {
+                                    _model.textController?.clear();
+                                  });
+                                },
+                                child: Icon(
+                                  Icons.send_rounded,
+                                  color: FlutterFlowTheme.of(context).main1,
+                                  size: 25.0,
+                                ),
                               ),
                             ),
                           ],
