@@ -2,11 +2,11 @@ import '/backend/backend.dart';
 import '/flutter_flow/flutter_flow_theme.dart';
 import '/flutter_flow/flutter_flow_toggle_icon.dart';
 import '/flutter_flow/flutter_flow_util.dart';
-import '/flutter_flow/flutter_flow_widgets.dart';
+import '/flutter_flow/custom_functions.dart' as functions;
 import 'package:smooth_page_indicator/smooth_page_indicator.dart'
     as smooth_page_indicator;
-import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
@@ -15,9 +15,9 @@ export 'post_detail_page_model.dart';
 
 class PostDetailPageWidget extends StatefulWidget {
   const PostDetailPageWidget({
-    Key? key,
+    super.key,
     required this.postRef,
-  }) : super(key: key);
+  });
 
   final DocumentReference? postRef;
 
@@ -47,6 +47,15 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
 
   @override
   Widget build(BuildContext context) {
+    if (isiOS) {
+      SystemChrome.setSystemUIOverlayStyle(
+        SystemUiOverlayStyle(
+          statusBarBrightness: Theme.of(context).brightness,
+          systemStatusBarContrastEnforced: true,
+        ),
+      );
+    }
+
     context.watch<FFAppState>();
 
     return StreamBuilder<PostRecord>(
@@ -55,7 +64,6 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
         // Customize what your widget looks like when it's loading.
         if (!snapshot.hasData) {
           return Scaffold(
-            backgroundColor: FlutterFlowTheme.of(context).primaryBackground,
             body: Center(
               child: SizedBox(
                 width: 50.0,
@@ -70,17 +78,20 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
         }
         final postDetailPagePostRecord = snapshot.data!;
         return GestureDetector(
-          onTap: () => FocusScope.of(context).requestFocus(_model.unfocusNode),
+          onTap: () => _model.unfocusNode.canRequestFocus
+              ? FocusScope.of(context).requestFocus(_model.unfocusNode)
+              : FocusScope.of(context).unfocus(),
           child: Scaffold(
             key: scaffoldKey,
-            backgroundColor: FlutterFlowTheme.of(context).primaryBackground,
             body: SafeArea(
               top: true,
-              child: Stack(
-                children: [
-                  SingleChildScrollView(
-                    child: Column(
+              child: Align(
+                alignment: const AlignmentDirectional(0.0, 0.0),
+                child: Stack(
+                  children: [
+                    Column(
                       mainAxisSize: MainAxisSize.max,
+                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
                       children: [
                         Container(
                           width: MediaQuery.sizeOf(context).width * 1.0,
@@ -96,7 +107,7 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                   final images = postDetailPagePostRecord
                                       .uploadedImages
                                       .toList();
-                                  return Container(
+                                  return SizedBox(
                                     width: double.infinity,
                                     height: 500.0,
                                     child: Stack(
@@ -112,24 +123,44 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                           itemBuilder: (context, imagesIndex) {
                                             final imagesItem =
                                                 images[imagesIndex];
-                                            return ClipRRect(
-                                              borderRadius:
-                                                  BorderRadius.circular(0.0),
-                                              child: Image.network(
-                                                imagesItem,
-                                                width: 300.0,
-                                                height: 200.0,
-                                                fit: BoxFit.cover,
+                                            return InkWell(
+                                              splashColor: Colors.transparent,
+                                              focusColor: Colors.transparent,
+                                              hoverColor: Colors.transparent,
+                                              highlightColor:
+                                                  Colors.transparent,
+                                              onTap: () async {
+                                                context.pushNamed(
+                                                  'showImage',
+                                                  queryParameters: {
+                                                    'postDocument':
+                                                        serializeParam(
+                                                      widget.postRef,
+                                                      ParamType
+                                                          .DocumentReference,
+                                                    ),
+                                                  }.withoutNulls,
+                                                );
+                                              },
+                                              child: ClipRRect(
+                                                borderRadius:
+                                                    BorderRadius.circular(0.0),
+                                                child: Image.network(
+                                                  imagesItem,
+                                                  width: 300.0,
+                                                  height: 200.0,
+                                                  fit: BoxFit.cover,
+                                                ),
                                               ),
                                             );
                                           },
                                         ),
                                         Align(
                                           alignment:
-                                              AlignmentDirectional(0.00, 1.00),
+                                              const AlignmentDirectional(0.0, 1.0),
                                           child: Padding(
                                             padding:
-                                                EdgeInsetsDirectional.fromSTEB(
+                                                const EdgeInsetsDirectional.fromSTEB(
                                                     0.0, 0.0, 0.0, 18.0),
                                             child: smooth_page_indicator
                                                 .SmoothPageIndicator(
@@ -144,7 +175,7 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                                 await _model.pageViewController!
                                                     .animateToPage(
                                                   i,
-                                                  duration: Duration(
+                                                  duration: const Duration(
                                                       milliseconds: 500),
                                                   curve: Curves.ease,
                                                 );
@@ -175,9 +206,9 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                 width: MediaQuery.sizeOf(context).width * 1.0,
                                 height:
                                     MediaQuery.sizeOf(context).height * 0.08,
-                                decoration: BoxDecoration(),
+                                decoration: const BoxDecoration(),
                                 child: Padding(
-                                  padding: EdgeInsetsDirectional.fromSTEB(
+                                  padding: const EdgeInsetsDirectional.fromSTEB(
                                       20.0, 0.0, 20.0, 0.0),
                                   child: Row(
                                     mainAxisSize: MainAxisSize.max,
@@ -197,13 +228,13 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                             onTap: () async {
                                               context.safePop();
                                             },
-                                            child: Icon(
+                                            child: const Icon(
                                               Icons.arrow_back_ios,
                                               color: Colors.white,
                                               size: 24.0,
                                             ),
                                           ),
-                                          Padding(
+                                          const Padding(
                                             padding:
                                                 EdgeInsetsDirectional.fromSTEB(
                                                     10.0, 0.0, 0.0, 0.0),
@@ -215,7 +246,7 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                           ),
                                         ],
                                       ),
-                                      Row(
+                                      const Row(
                                         mainAxisSize: MainAxisSize.min,
                                         mainAxisAlignment:
                                             MainAxisAlignment.spaceBetween,
@@ -265,9 +296,11 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                             return Container(
                               width: MediaQuery.sizeOf(context).width * 1.0,
                               height: MediaQuery.sizeOf(context).height * 0.55,
-                              decoration: BoxDecoration(),
+                              decoration: const BoxDecoration(),
                               child: Column(
                                 mainAxisSize: MainAxisSize.max,
+                                mainAxisAlignment:
+                                    MainAxisAlignment.spaceBetween,
                                 children: [
                                   StreamBuilder<UserRecord>(
                                     stream: UserRecord.getDocument(
@@ -297,19 +330,19 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                         height:
                                             MediaQuery.sizeOf(context).height *
                                                 0.1,
-                                        decoration: BoxDecoration(),
+                                        decoration: const BoxDecoration(),
                                         child: Row(
                                           mainAxisSize: MainAxisSize.max,
                                           children: [
                                             Padding(
-                                              padding: EdgeInsetsDirectional
+                                              padding: const EdgeInsetsDirectional
                                                   .fromSTEB(
                                                       15.0, 0.0, 0.0, 0.0),
                                               child: Container(
                                                 width: 55.0,
                                                 height: 55.0,
                                                 clipBehavior: Clip.antiAlias,
-                                                decoration: BoxDecoration(
+                                                decoration: const BoxDecoration(
                                                   shape: BoxShape.circle,
                                                 ),
                                                 child: Image.network(
@@ -319,7 +352,7 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                               ),
                                             ),
                                             Padding(
-                                              padding: EdgeInsetsDirectional
+                                              padding: const EdgeInsetsDirectional
                                                   .fromSTEB(
                                                       10.0, 0.0, 0.0, 0.0),
                                               child: Column(
@@ -331,7 +364,7 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                                 children: [
                                                   Padding(
                                                     padding:
-                                                        EdgeInsetsDirectional
+                                                        const EdgeInsetsDirectional
                                                             .fromSTEB(0.0, 0.0,
                                                                 0.0, 10.0),
                                                     child: Text(
@@ -357,7 +390,7 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                                     ),
                                                   ),
                                                   Text(
-                                                    'something level',
+                                                    'User address',
                                                     style: FlutterFlowTheme.of(
                                                             context)
                                                         .bodyMedium
@@ -388,50 +421,79 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                     width:
                                         MediaQuery.sizeOf(context).width * 1.0,
                                     height: MediaQuery.sizeOf(context).height *
-                                        0.35,
-                                    decoration: BoxDecoration(),
+                                        0.29,
+                                    decoration: const BoxDecoration(),
                                     child: Padding(
-                                      padding: EdgeInsetsDirectional.fromSTEB(
-                                          15.0, 5.0, 15.0, 0.0),
-                                      child: Column(
-                                        mainAxisSize: MainAxisSize.max,
-                                        crossAxisAlignment:
-                                            CrossAxisAlignment.start,
-                                        children: [
-                                          Divider(
-                                            thickness: 1.0,
-                                            color: Color(0xFFB9B0B0),
-                                          ),
-                                          Text(
-                                            postDetailPagePostRecord.title,
-                                            style: FlutterFlowTheme.of(context)
-                                                .titleLarge
-                                                .override(
-                                                  fontFamily:
-                                                      FlutterFlowTheme.of(
-                                                              context)
-                                                          .titleLargeFamily,
-                                                  fontSize: 22.0,
-                                                  useGoogleFonts: GoogleFonts
-                                                          .asMap()
-                                                      .containsKey(
-                                                          FlutterFlowTheme.of(
-                                                                  context)
-                                                              .titleLargeFamily),
+                                      padding: const EdgeInsetsDirectional.fromSTEB(
+                                          15.0, 0.0, 15.0, 0.0),
+                                      child: SingleChildScrollView(
+                                        child: Column(
+                                          mainAxisSize: MainAxisSize.max,
+                                          crossAxisAlignment:
+                                              CrossAxisAlignment.start,
+                                          children: [
+                                            const Divider(
+                                              thickness: 1.0,
+                                              color: Color(0xFFB9B0B0),
+                                            ),
+                                            Text(
+                                              postDetailPagePostRecord.title,
+                                              style:
+                                                  FlutterFlowTheme.of(context)
+                                                      .titleLarge
+                                                      .override(
+                                                        fontFamily:
+                                                            FlutterFlowTheme.of(
+                                                                    context)
+                                                                .titleLargeFamily,
+                                                        fontSize: 22.0,
+                                                        useGoogleFonts: GoogleFonts
+                                                                .asMap()
+                                                            .containsKey(
+                                                                FlutterFlowTheme.of(
+                                                                        context)
+                                                                    .titleLargeFamily),
+                                                      ),
+                                            ),
+                                            Row(
+                                              mainAxisSize: MainAxisSize.max,
+                                              mainAxisAlignment:
+                                                  MainAxisAlignment.start,
+                                              children: [
+                                                Padding(
+                                                  padding: const EdgeInsetsDirectional
+                                                      .fromSTEB(
+                                                          0.0, 0.0, 20.0, 0.0),
+                                                  child: Text(
+                                                    postDetailPagePostRecord
+                                                        .category,
+                                                    style: FlutterFlowTheme.of(
+                                                            context)
+                                                        .labelSmall
+                                                        .override(
+                                                          fontFamily:
+                                                              FlutterFlowTheme.of(
+                                                                      context)
+                                                                  .labelSmallFamily,
+                                                          color:
+                                                              const Color(0xFF7A7878),
+                                                          decoration:
+                                                              TextDecoration
+                                                                  .underline,
+                                                          useGoogleFonts: GoogleFonts
+                                                                  .asMap()
+                                                              .containsKey(
+                                                                  FlutterFlowTheme.of(
+                                                                          context)
+                                                                      .labelSmallFamily),
+                                                        ),
+                                                  ),
                                                 ),
-                                          ),
-                                          Row(
-                                            mainAxisSize: MainAxisSize.max,
-                                            mainAxisAlignment:
-                                                MainAxisAlignment.start,
-                                            children: [
-                                              Padding(
-                                                padding: EdgeInsetsDirectional
-                                                    .fromSTEB(
-                                                        0.0, 0.0, 20.0, 0.0),
-                                                child: Text(
-                                                  postDetailPagePostRecord
-                                                      .category,
+                                                Text(
+                                                  dateTimeFormat(
+                                                      'jm',
+                                                      postDetailPagePostRecord
+                                                          .uploadTime!),
                                                   style: FlutterFlowTheme.of(
                                                           context)
                                                       .labelSmall
@@ -441,10 +503,7 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                                                     context)
                                                                 .labelSmallFamily,
                                                         color:
-                                                            Color(0xFF7A7878),
-                                                        decoration:
-                                                            TextDecoration
-                                                                .underline,
+                                                            const Color(0xFF7A7878),
                                                         useGoogleFonts: GoogleFonts
                                                                 .asMap()
                                                             .containsKey(
@@ -453,12 +512,15 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                                                     .labelSmallFamily),
                                                       ),
                                                 ),
-                                              ),
-                                              Text(
-                                                dateTimeFormat(
-                                                    'jm',
-                                                    postDetailPagePostRecord
-                                                        .uploadTime!),
+                                              ],
+                                            ),
+                                            Padding(
+                                              padding: const EdgeInsetsDirectional
+                                                  .fromSTEB(
+                                                      0.0, 25.0, 0.0, 0.0),
+                                              child: Text(
+                                                postDetailPagePostRecord
+                                                    .contents,
                                                 style:
                                                     FlutterFlowTheme.of(context)
                                                         .labelSmall
@@ -467,8 +529,7 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                                               FlutterFlowTheme.of(
                                                                       context)
                                                                   .labelSmallFamily,
-                                                          color:
-                                                              Color(0xFF7A7878),
+                                                          fontSize: 14.0,
                                                           useGoogleFonts: GoogleFonts
                                                                   .asMap()
                                                               .containsKey(
@@ -477,184 +538,187 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                                                                       .labelSmallFamily),
                                                         ),
                                               ),
-                                            ],
-                                          ),
-                                          Padding(
-                                            padding:
-                                                EdgeInsetsDirectional.fromSTEB(
-                                                    0.0, 25.0, 0.0, 0.0),
-                                            child: Text(
-                                              postDetailPagePostRecord.contents,
-                                              style:
-                                                  FlutterFlowTheme.of(context)
-                                                      .labelSmall
-                                                      .override(
-                                                        fontFamily:
-                                                            FlutterFlowTheme.of(
-                                                                    context)
-                                                                .labelSmallFamily,
-                                                        fontSize: 14.0,
-                                                        useGoogleFonts: GoogleFonts
-                                                                .asMap()
-                                                            .containsKey(
-                                                                FlutterFlowTheme.of(
-                                                                        context)
-                                                                    .labelSmallFamily),
-                                                      ),
-                                            ),
-                                          ),
-                                        ],
-                                      ),
-                                    ),
-                                  ),
-                                  Container(
-                                    width:
-                                        MediaQuery.sizeOf(context).width * 1.0,
-                                    height: MediaQuery.sizeOf(context).height *
-                                        0.08,
-                                    decoration: BoxDecoration(),
-                                    child: Row(
-                                      mainAxisSize: MainAxisSize.max,
-                                      mainAxisAlignment:
-                                          MainAxisAlignment.spaceEvenly,
-                                      children: [
-                                        Padding(
-                                          padding:
-                                              EdgeInsetsDirectional.fromSTEB(
-                                                  10.0, 0.0, 0.0, 0.0),
-                                          child: ToggleIcon(
-                                            onPressed: () async {
-                                              setState(() =>
-                                                  FFAppState().likeToggle =
-                                                      !FFAppState().likeToggle);
-                                            },
-                                            value: FFAppState().likeToggle,
-                                            onIcon: Icon(
-                                              Icons.favorite,
-                                              color:
-                                                  FlutterFlowTheme.of(context)
-                                                      .alternate,
-                                              size: 25.0,
-                                            ),
-                                            offIcon: Icon(
-                                              Icons.favorite_border_sharp,
-                                              color:
-                                                  FlutterFlowTheme.of(context)
-                                                      .sub1,
-                                              size: 25.0,
-                                            ),
-                                          ),
-                                        ),
-                                        SizedBox(
-                                          height: 100.0,
-                                          child: VerticalDivider(
-                                            thickness: 1.0,
-                                            color: Color(0xFFB8B4B4),
-                                          ),
-                                        ),
-                                        Column(
-                                          mainAxisSize: MainAxisSize.max,
-                                          mainAxisAlignment:
-                                              MainAxisAlignment.center,
-                                          crossAxisAlignment:
-                                              CrossAxisAlignment.start,
-                                          children: [
-                                            Text(
-                                              'price',
-                                              style:
-                                                  FlutterFlowTheme.of(context)
-                                                      .headlineSmall,
-                                            ),
-                                            Text(
-                                              'Fixed Price',
-                                              style:
-                                                  FlutterFlowTheme.of(context)
-                                                      .titleSmall
-                                                      .override(
-                                                        fontFamily:
-                                                            FlutterFlowTheme.of(
-                                                                    context)
-                                                                .titleSmallFamily,
-                                                        color:
-                                                            Color(0xC2C0BCBC),
-                                                        useGoogleFonts: GoogleFonts
-                                                                .asMap()
-                                                            .containsKey(
-                                                                FlutterFlowTheme.of(
-                                                                        context)
-                                                                    .titleSmallFamily),
-                                                      ),
                                             ),
                                           ],
                                         ),
-                                        Align(
-                                          alignment:
-                                              AlignmentDirectional(0.00, 0.00),
-                                          child: Padding(
+                                      ),
+                                    ),
+                                  ),
+                                  Align(
+                                    alignment: const AlignmentDirectional(0.0, 0.0),
+                                    child: Container(
+                                      width: MediaQuery.sizeOf(context).width *
+                                          1.0,
+                                      height:
+                                          MediaQuery.sizeOf(context).height *
+                                              0.08,
+                                      decoration: const BoxDecoration(),
+                                      child: Row(
+                                        mainAxisSize: MainAxisSize.max,
+                                        mainAxisAlignment:
+                                            MainAxisAlignment.spaceEvenly,
+                                        crossAxisAlignment:
+                                            CrossAxisAlignment.center,
+                                        children: [
+                                          Padding(
                                             padding:
-                                                EdgeInsetsDirectional.fromSTEB(
-                                                    100.0, 0.0, 0.0, 0.0),
-                                            child: InkWell(
-                                              splashColor: Colors.transparent,
-                                              focusColor: Colors.transparent,
-                                              hoverColor: Colors.transparent,
-                                              highlightColor:
-                                                  Colors.transparent,
-                                              onTap: () async {
-                                                context.pushNamed(
-                                                  'ChatPage',
-                                                  queryParameters: {
-                                                    'chatUser': serializeParam(
-                                                      containerUserRecord,
-                                                      ParamType.Document,
-                                                    ),
-                                                  }.withoutNulls,
-                                                  extra: <String, dynamic>{
-                                                    'chatUser':
-                                                        containerUserRecord,
-                                                  },
-                                                );
+                                                const EdgeInsetsDirectional.fromSTEB(
+                                                    10.0, 0.0, 0.0, 0.0),
+                                            child: ToggleIcon(
+                                              onPressed: () async {
+                                                setState(() => FFAppState()
+                                                        .likeToggle =
+                                                    !FFAppState().likeToggle);
                                               },
-                                              child: Container(
-                                                width: 80.0,
-                                                height: 50.0,
-                                                decoration: BoxDecoration(
-                                                  color: FlutterFlowTheme.of(
-                                                          context)
-                                                      .tertiary,
-                                                  borderRadius:
-                                                      BorderRadius.circular(
-                                                          12.0),
-                                                ),
-                                                child: Align(
-                                                  alignment:
-                                                      AlignmentDirectional(
-                                                          0.00, 0.00),
-                                                  child: Text(
-                                                    'Chat',
-                                                    style: FlutterFlowTheme.of(
-                                                            context)
-                                                        .titleMedium
+                                              value: FFAppState().likeToggle,
+                                              onIcon: Icon(
+                                                Icons.favorite,
+                                                color:
+                                                    FlutterFlowTheme.of(context)
+                                                        .alternate,
+                                                size: 25.0,
+                                              ),
+                                              offIcon: Icon(
+                                                Icons.favorite_border_sharp,
+                                                color:
+                                                    FlutterFlowTheme.of(context)
+                                                        .sub1,
+                                                size: 25.0,
+                                              ),
+                                            ),
+                                          ),
+                                          const SizedBox(
+                                            height: 40.0,
+                                            child: VerticalDivider(
+                                              thickness: 1.0,
+                                              color: Color(0xFFB8B4B4),
+                                            ),
+                                          ),
+                                          Column(
+                                            mainAxisSize: MainAxisSize.max,
+                                            mainAxisAlignment:
+                                                MainAxisAlignment.center,
+                                            crossAxisAlignment:
+                                                CrossAxisAlignment.start,
+                                            children: [
+                                              Text(
+                                                '${postDetailPagePostRecord.price.toString()}원',
+                                                style:
+                                                    FlutterFlowTheme.of(context)
+                                                        .headlineSmall
                                                         .override(
                                                           fontFamily:
                                                               FlutterFlowTheme.of(
                                                                       context)
-                                                                  .titleMediumFamily,
-                                                          color: Colors.white,
+                                                                  .headlineSmallFamily,
+                                                          fontSize: 18.0,
                                                           useGoogleFonts: GoogleFonts
                                                                   .asMap()
                                                               .containsKey(
                                                                   FlutterFlowTheme.of(
                                                                           context)
-                                                                      .titleMediumFamily),
+                                                                      .headlineSmallFamily),
                                                         ),
+                                              ),
+                                              Text(
+                                                valueOrDefault<String>(
+                                                  functions.checkNegoitable(
+                                                      postDetailPagePostRecord
+                                                          .negoitable),
+                                                  'false',
+                                                ),
+                                                style:
+                                                    FlutterFlowTheme.of(context)
+                                                        .titleSmall
+                                                        .override(
+                                                          fontFamily:
+                                                              FlutterFlowTheme.of(
+                                                                      context)
+                                                                  .titleSmallFamily,
+                                                          color:
+                                                              const Color(0xC2C0BCBC),
+                                                          fontSize: 14.0,
+                                                          useGoogleFonts: GoogleFonts
+                                                                  .asMap()
+                                                              .containsKey(
+                                                                  FlutterFlowTheme.of(
+                                                                          context)
+                                                                      .titleSmallFamily),
+                                                        ),
+                                              ),
+                                            ],
+                                          ),
+                                          Align(
+                                            alignment:
+                                                const AlignmentDirectional(0.0, 0.0),
+                                            child: Padding(
+                                              padding: const EdgeInsetsDirectional
+                                                  .fromSTEB(
+                                                      100.0, 0.0, 0.0, 0.0),
+                                              child: InkWell(
+                                                splashColor: Colors.transparent,
+                                                focusColor: Colors.transparent,
+                                                hoverColor: Colors.transparent,
+                                                highlightColor:
+                                                    Colors.transparent,
+                                                onTap: () async {
+                                                  context.pushNamed(
+                                                    'ChatPage',
+                                                    queryParameters: {
+                                                      'chatUser':
+                                                          serializeParam(
+                                                        containerUserRecord,
+                                                        ParamType.Document,
+                                                      ),
+                                                    }.withoutNulls,
+                                                    extra: <String, dynamic>{
+                                                      'chatUser':
+                                                          containerUserRecord,
+                                                    },
+                                                  );
+                                                },
+                                                child: Container(
+                                                  width: 80.0,
+                                                  height: 40.0,
+                                                  decoration: BoxDecoration(
+                                                    color: FlutterFlowTheme.of(
+                                                            context)
+                                                        .tertiary,
+                                                    borderRadius:
+                                                        BorderRadius.circular(
+                                                            12.0),
+                                                  ),
+                                                  child: Align(
+                                                    alignment:
+                                                        const AlignmentDirectional(
+                                                            0.0, 0.0),
+                                                    child: Text(
+                                                      'Chat',
+                                                      style:
+                                                          FlutterFlowTheme.of(
+                                                                  context)
+                                                              .titleMedium
+                                                              .override(
+                                                                fontFamily: FlutterFlowTheme.of(
+                                                                        context)
+                                                                    .titleMediumFamily,
+                                                                color: Colors
+                                                                    .white,
+                                                                fontSize: 16.0,
+                                                                useGoogleFonts: GoogleFonts
+                                                                        .asMap()
+                                                                    .containsKey(
+                                                                        FlutterFlowTheme.of(context)
+                                                                            .titleMediumFamily),
+                                                              ),
+                                                    ),
                                                   ),
                                                 ),
                                               ),
                                             ),
                                           ),
-                                        ),
-                                      ],
+                                        ],
+                                      ),
                                     ),
                                   ),
                                 ],
@@ -664,8 +728,8 @@ class _PostDetailPageWidgetState extends State<PostDetailPageWidget> {
                         ),
                       ],
                     ),
-                  ),
-                ],
+                  ],
+                ),
               ),
             ),
           ),
